% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{analysis}
\alias{analysis}
\title{comparison of two RGCCA results}
\usage{
analysis(refData, noIntersect = TRUE, C = NULL, tau = NULL,
  scale = TRUE, nAxe = 2, scheme = "centroid",
  sameBlockWeight = TRUE, wd = getwd(), blocknames = NULL,
  nbTestFiles = 20)
}
\arguments{
\item{refData}{A list of complete blocks}

\item{noIntersect}{if TRUE, no RGCCA on complete data is run}

\item{C, tau, scale, scheme, sameBlockWeight}{parameters of RGCCA}
}
\value{
\item{A}{A list containing: a: the correlation between axes, rv: the rv coefficient, bm the biomarkers}

\item{crit}{Convergence criterion : abs(1-obj_k/obj_{k-1})}

\item{obj}{Vector containing the mean square error between the predict values and the original non missing values at each iteration}
}
\description{
Analysis of the comparison of different NA methods on RGCCA
}
\examples{
 data();...
}
