% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{available_methods}
\alias{available_methods}
\title{Available methods for RGCCA}
\usage{
available_methods()
}
\value{
A vector of the methods implemented with the rgcca function.
}
\description{
List the methods that can be used with the rgcca function.
}
