% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySearch.R
\name{BinarySearch}
\alias{BinarySearch}
\title{Internal function which does not have to be used by the users}
\usage{
BinarySearch(argu, sumabs)
}
\description{
Internal function which does not have to be used by the users
}
\author{
Arnaud Gloaguen
}
\keyword{internal}
