% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap confidence intervals and p-values}
\usage{
bootstrap(
  rgcca_res,
  n_boot = 100,
  n_cores = parallel::detectCores() - 1,
  parallelization = NULL
)
}
\arguments{
\item{rgcca_res}{A RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{n_boot}{Number of bootstrap iterations. Default is 100.}

\item{n_cores}{Number of cores for parallelization}

\item{parallelization}{if TRUE, the bootstrap is processed in parallel. If 
parallelization = NULL (default), parallelization is always performed except 
for Windows if length(n_boot) < 10.}
}
\value{
A list containing two objects: 'bootstrap' and 'rgcca'. 
'bootstrap' is a list containing for each block, a matrix 
with the variables of the block in row and the block weight vector
calculated accross bootstrap sample in column. 'rgcca' is the original 
fitted rgcca object. (see  \code{\link[RGCCA]{rgcca}})
}
\description{
Boostrap confidence intervals and p-values for evaluating the significancy/ 
stability of the block-weight vectors produce by S/RGCCA.
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], 
              industry = Russett[, 4:5], 
              politic = Russett[, c(6:9, 11)]
              )
              
fit.rgcca = rgcca(blocks)
boot_rgcca = bootstrap(fit.rgcca, n_boot = 50)
plot(boot_rgcca, block = 3)

}
\seealso{
\code{\link[RGCCA]{plot.bootstrap}} , 
\code{\link[RGCCA]{print.bootstrap}}
}
