% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_list.R
\name{char_to_list}
\alias{char_to_list}
\title{Convert a character in a vector}
\usage{
char_to_list(x)
}
\arguments{
\item{x}{A character separated by comma}
}
\value{
A vector of characters whitout spaces
}
\description{
Convert a character in a vector
}
\examples{
s = '1,2, 3'
char_to_list(s)
}
