% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_connection.R
\name{check_connection}
\alias{check_connection}
\title{Check the format of the connection matrix}
\usage{
check_connection(c, blocks)
}
\arguments{
\item{c}{A symmetric matrix containing 1 and 0}

\item{blocks}{A list of matrix}
}
\description{
Check the format of the connection matrix
}
