% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quantitative.R
\name{check_quantitative}
\alias{check_quantitative}
\title{Check if a dataframe contains no qualitative variables}
\usage{
check_quantitative(df, fo, header = FALSE, warn_separator = FALSE)
}
\arguments{
\item{df}{A dataframe or a matrix}

\item{fo}{A character giving the name of the tested file}

\item{header}{A bolean giving either the presence (TRUE) or absence (FALSE) of
a header}

\item{warn_separator}{A bolean to print warning for bad separator use}
}
\description{
Check if a dataframe contains no qualitative variables
}
