% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{checkQuantitative}
\alias{checkQuantitative}
\title{Check if a dataframe contains no quanlitative variables}
\usage{
checkQuantitative(df, fo, h = FALSE)
}
\arguments{
\item{df}{A dataframe or a matrix}

\item{fo}{A character giving the name of the tested file}

\item{h}{A bolean giving either the presence (TRUE) or absence (FALSE) of a header}
}
\description{
Check if a dataframe contains no quanlitative variables
}
\examples{
df = matrix(runif(20), 10, 2)
checkQuantitative(df, "data")
\dontrun{
df[,2] = LETTERS[1:10]
checkQuantitative(df, "data", TRUE)
# Error
}
}
