% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_group.R
\name{color_group}
\alias{color_group}
\title{Groups of color}
\usage{
color_group(x, colors = NULL)
}
\arguments{
\item{x}{A vector}
}
\value{
A color vector of equal size to the input vector
}
\description{
Returns a color vector of equal size to the input vector
}
\examples{
color_group(seq(10))
}
