% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_rows.r
\name{common_rows}
\alias{common_rows}
\title{Keep only the rows with the same names among a list of dataframe}
\usage{
common_rows(list_m)
}
\arguments{
\item{list_m}{A list of matrix}
}
\value{
A list of matrix
}
\description{
Keep only the rows with the same names among a list of dataframe
}
