% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{commonRow}
\alias{commonRow}
\title{Get the rows with the same names among a list of dataframe}
\usage{
commonRow(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A vector of character with the common rownames
}
\description{
Get the rows with the same names among a list of dataframe
}
