% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{correct}
\alias{correct}
\title{Check that the network is fully connected}
\usage{
correct(x)
}
\arguments{
\item{x}{Design matrix, a symmetric matrix with}
}
\value{
A logical value if it is fully connected or not.
}
\description{
Given the design matrix, checks that all the blocks are connected between them
}
\examples{
C <- matrix(c(0, 1, 0, 1, 0, 0, 0, 0, 0), ncol = 3, nrow = 3)
correct(C)
}
\references{
\url{https://math.stackexchange.com/a/551947}
}
\author{
Lluís Revilla Sancho
}
