% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov3.R
\name{cov3}
\alias{cov3}
\title{cov3}
\usage{
cov3(x, y = NULL, bias = TRUE)
}
\arguments{
\item{x}{a matrix x}

\item{y}{a matrix with same size as x}

\item{bias}{if TRUE, the estimator of variance is SS/sqrt(n-1), if FALSE, it is SS/sqrt(n)}
}
\description{
Calculates the covariance on the available data
}
\examples{
library(RGCCA)
X1=matrix(rnorm(35),7,5);
X2=matrix(rnorm(35),7,5);
# Check old functionalities
cov3(X1,bias=FALSE)==cov(X1)
cov3(X1,X2,bias=FALSE)==cov(X1,X2)
}
