% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivation.R
\name{derivation}
\alias{derivation}
\title{compute the derivative of the g function}
\usage{
derivation(g)
}
\arguments{
\item{g}{a convex function}
}
\value{
\item{dg}{the derivate of the g function}
}
\description{
The function derivation() allows computing the derivative of the g function. 
The g function must be convex.
}

