% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_patternNA.R
\name{determine_patternNA}
\alias{determine_patternNA}
\title{determine_patternNA
Determines the pattern of missing values and allows to use it to create datasets with the same pattern.}
\usage{
determine_patternNA(A, graph = "all", legend = FALSE,
  outlierVisible = FALSE)
}
\arguments{
\item{A}{list of blocks}

\item{graph}{if TRUE, the pattern of missing values is plotted}
}
\value{
a list containing the percentage of missing values per variable (pctNA), the percentage of missing values per block (pctNAbyBlock), and the complete individuals percentage (completeSubjectByBlock)
}
\description{
determine_patternNA
Determines the pattern of missing values and allows to use it to create datasets with the same pattern.
}
\examples{
X1=matrix(rnorm(150),30,5)
X2=matrix(rnorm(150),30,5)
X3=matrix(rnorm(150),30,5)
X1[1:10,1]=NA
X1[2,2]=NA
X1[11,]=NA
X2[1:10,1]=NA
colnames(X1)=paste0("A",1:5)
colnames(X2)=paste0("B",1:5)
colnames(X3)=paste0("C",1:5)
A=list(bloc1=X1,bloc2=X2,bloc3=X3)
determine_patternNA(A)
}
