% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_confounders.R
\docType{data}
\name{example_confounders}
\alias{example_confounders}
\title{Biological data impacted by confounding variation:
Example data to illustrate use of AC-RGCCA}
\format{
A list of 6 elements
}
\usage{
data(example_confounders)
}
\description{
This simulated dataset emulates a multi-modal data harmonization scenario, 
involving 
\eqn{I=5}{I=5} laboratories processing \eqn{N=10}{N=10} conditions, examined
via \eqn{J=2}{J=2} modalities (each observed through \eqn{p_j=500}{pj=500} 
variables with \eqn{j=1,2}{j=1,2}). The conditions represent varying doses
of exposure to a pollutant. Each laboratory produces a batch of \eqn{N}{N} 
samples, each one exposed to a different dose, creating a common dose effect
across batches that we aim to recover. The \eqn{J=2}{J=2} blocks of data are 
both simulated as the sum of three terms: a dose effect, simulated based on 
the same common pattern, a batch-specific term, and a random noise term. To 
model different types of unwanted effect, we assume a simple batch effect
on the first modality (which influences all samples from the same batch 
similarly and encourages them to cluster together) and a more complex 
confounding effect on the second modality, presenting a combined effect with
a batch effect that is not uniform across individuals. 

\describe{

\item{blocks}{A list of \eqn{J=2}{J=2} matrices called \code{block1} and 
\code{block2}, both with \eqn{n=NI=50}{n=NI=50} observations of 
\eqn{p_j=500}{pj=500} variables (\eqn{j=1,2}{j=1,2}). Both matrices were 
centered and scaled by columns.}

\item{confounders_Z_A}{A matrix of \eqn{n=NI=50}{n=NI=50} observations of 
\eqn{I=5}{I=5} variables defined as the disjunctive table encoding samples'
laboratory of origin. The matrix was centered and scaled by columns. Used 
as input for \code{confounders} in \code{rgcca}, this matrix will serve to 
compute the linear kernel matrix 
\eqn{\mathbf K = \mathbf Z_A \mathbf Z_A^\top}{K=Z_A Z_A'} and implement the
penalty scheme A, thus seeking components that are robust to the confounding 
variation.}

\item{confounders_Z_B}{A matrix of \eqn{n=NI=50}{n=NI=50} observations of 
\eqn{Q=NI(I-1)/2=100}{Q=NI(I-1)/2=100} variables used to match pairs of 
individuals within the same dose group. The matrix was centered and scaled 
by columns. Used as input for \code{confounders} in \code{rgcca}, this 
matrix will serve to compute the linear kernel matrix 
\eqn{\mathbf K = \mathbf Z_B \mathbf Z_B^\top}{K=Z_B Z_B'} and implement the
penalty scheme B, thus seeking components that allow to discriminate between
dose conditions, depite unwanted effects.}

\item{common_pattern}{A data frame of \eqn{N=10}{N=10} observations of 2 
variables giving the common pattern that was used to simulate the shared 
dose effect in the two blocks.}

\item{labs}{A factor of length \eqn{n=NI=50}{n=NI=50} giving the laboratory
of origin of a sample. Samples with the same \code{labs} value have been
produced in the same batch.}

\item{dose_groups}{A factor of length \eqn{n=NI=50}{n=NI=50} giving the
exposure group of all samples. Samples with the same \code{dose_groups} 
value are assumed to have received the same dose.}

}
}
\references{
Goujon E. et al. AC-RGCCA: Adjusting for Confounding Variation 
Within Multiblock Data Analysis. To be published.
}
\keyword{datasets}
