% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAback.R
\name{findAback}
\alias{findAback}
\title{findAback
finds the original values of A (scaled)}
\usage{
findAback(A, sameBlockWeight = TRUE)
}
\arguments{
\item{A}{results of scale}
}
\description{
findAback
finds the original values of A (scaled)
}
