% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_resampling.R
\name{generate_resampling}
\alias{generate_resampling}
\title{Generate sampled blocks for bootstrap.}
\usage{
generate_resampling(
  rgcca_res,
  n_boot,
  balanced = TRUE,
  keep_all_variables = FALSE,
  pval = 1e-15
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{n_boot}{Number of bootstrap samples. Default: 100.}

\item{balanced}{A boolean indicating if a balanced bootstrap procedure is
performed or not (default is TRUE).}

\item{keep_all_variables}{A boolean indicating if all variables have to be
kept even when some of them have null variance for at least one bootstrap
sample (default is FALSE).}

\item{pval}{For all the variables, a threshold for the proportion of the most
frequent value of this variable is computed. This threshold is evaluated so
that the probability to sample only this value is below `pval`. This
probability is corrected by the number of bootstrap samples and the number
of variables (default is `1e-15`).}
}
\value{
\item{full_idx}{A list of size n_boot containing the observations
kept for each bootstrap sample.}

\item{summarize_column_sd_null}{A list of size the number of block
containing the variables that were removed from each block for all the
bootstrap samples. Variables are removed if they appear to be of null variance
in at least one bootstrap sample. If no variable is removed, return NULL.}
}
\description{
Generate n_boot block-wise bootstrapped samples.
}
\keyword{internal}
