% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bloc_var.R
\name{get_bloc_var}
\alias{get_bloc_var}
\title{Get the blocs of each variables}
\usage{
get_bloc_var(df, collapse = FALSE)
}
\arguments{
\item{df}{A list of matrix where their names are those of the blocks and the 
superblock and their rows are named after their variables}

\item{collapse}{A boolean to combine the variables of each blocks as result}
}
\value{
A vector of character giving block names for each corresponding 
variable.
}
\description{
Get a vector of block names for each corresponding variable. The last block 
is considered as the superblock and ignored.
}
\seealso{
\code{\link[RGCCA]{rgccad}}, \code{\link[RGCCA]{sgcca}}
}
