% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap.R
\name{get_bootstrap}
\alias{get_bootstrap}
\title{Extract a bootstrap}
\usage{
get_bootstrap(b, comp = 1, i_block = length(b$bootstrap[[1]]),
  bars = "sd", collapse = FALSE, n_cores = parallel::detectCores() -
  1, display_order = TRUE)
}
\arguments{
\item{b}{A bootstrap object (see  \code{\link[RGCCA]{bootstrap}} )}

\item{comp}{An integer giving the index of the analysis components}

\item{i_block}{An integer giving the index of a list of blocks}

\item{bars}{A character among "sd" for standard deviations, "stderr" for the standard error, "ci" for confidence interval of scores and "cim" for the confidence intervall of the mean.}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{display_order}{A logical value to display the order of the variables}
}
\value{
A matrix containing the means, 95\% intervals, bootstrap ratio, p-values and other statistics (see details)
}
\description{
Extract statistical information from a bootstrap
}
\details{
\itemize{
\item 'mean' for the mean of the bootstrap weights
\item 'estimate' for RGCCA weights
\item 'sd' for the standard error of the bootstrap weights
\item 'lower/upper_band' for the lower and upper intervals from to the 'bar' parameter
\item 'bootstrap_ratio' for the mean of the bootstrap weights / their standard error
\item 'p.vals' for p-values. The occurrences of  are distributed in accordance with the law
\item 'BH' for Benjamini-Hochberg p-value adjustments
\item 'occurrences' for non-zero occurences (for SGCCA) 
\item 'sign' for significant 95% bootstrap interval (for RGCCA)
}
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, 5, n_cores = 1)
get_bootstrap(boot, n_cores = 1)
}
\seealso{
\code{\link[RGCCA]{bootstrap}}, \code{\link[RGCCA]{plot.bootstrap}} , \code{\link[RGCCA]{print.bootstrap}}
}
