% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp.R
\name{get_comp}
\alias{get_comp}
\title{Get the components of the analysis}
\usage{
get_comp(rgcca, resp = rep(1, NROW(rgcca$Y[[1]])), compx = 1,
  compy = 2, compz = NULL, i_block = length(rgcca$Y),
  i_block_y = i_block, i_block_z = i_block, predicted = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{compz}{An integer giving the index of the analysis component used
for the z-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{i_block_z}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{predicted}{A list containing as  2nd element a matrix of predicted components}
}
\value{
A matrix containg each selected components and an associated response
}
\description{
Get the components of the analysis
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks)
response = factor( apply(Russett[, 9:11], 1, which.max),
                  labels = colnames(Russett)[9:11] )
get_comp(rgcca_out, as.matrix(response))
response = as.matrix(runif(NROW(blocks[[1]])))
row.names(response) = row.names(blocks[[1]])
get_comp(rgcca_out, response)
}
