% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_all.R
\name{get_cor_all}
\alias{get_cor_all}
\title{Correlation between the blocks}
\usage{
get_cor_all(rgcca_res, blocks = rgcca_res$call$blocks,
  comps = get_comp_all(rgcca_res))
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{blocks}{A list of matrix}

\item{comps}{A matrix containg the components of all the blocks}
}
\value{
A list of matrix containg the pairwise correlation of the blocks
}
\description{
Calculates the pairwise correlation between each block for each of the 
components of the analysis
}
