% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctr.R
\name{get_ctr}
\alias{get_ctr}
\title{Variable contribution}
\usage{
get_ctr(rgcca, compx = 1, compy = 2, compz = NULL,
  i_block = length(rgcca$call$blocks), type = "cor",
  collapse = FALSE)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{compz}{An integer giving the index of the analysis component used
for the z-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A character giving the choice ot the index between cor or weight}

\item{collapse}{A boolean to combine the variables of each blocks as result}
}
\value{
A dataframe containing the indexes for each selected components
}
\description{
Extract the contibution of variables to the model by using correlation or weight
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks, ncomp = c(3,2,4))
get_ctr(rgcca_out)
# On the first block and with weights
get_ctr(rgcca_out, 2, 1, i_block = 1, type = "weight")
# With 3 components and on the variables of two blocks
superblocks <- rep(list(Reduce(cbind, c(blocks[1], blocks[3]))), 2)
names(superblocks) <- names(blocks)[c(1, 3)]
rgcca_out = rgcca.analyze(blocks[c(1,3)], ncomp = c(3,4))
rgcca_out$blocks = superblocks
get_ctr(rgcca_out, compz = 3, i_block = 1, type = "cor", collapse = TRUE)
get_ctr(rgcca_out, 2, 1, 3, 1, "weight", TRUE)
}
