% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename.R
\name{get_filename}
\alias{get_filename}
\title{File name from a path}
\usage{
get_filename(fi)
}
\arguments{
\item{fi}{A character giving the path of a file}
}
\value{
A character for the name of the file
}
\description{
Get the file name from a path
}
