% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getBlocsVariables}
\alias{getBlocsVariables}
\title{Get the blocs of each variables}
\usage{
getBlocsVariables(df)
}
\arguments{
\item{df}{A list of matrix where their names are those of the blocks and the superblock and their rows are named after their variables}
}
\value{
A vector of character giving block names for each corresponding variable.
}
\description{
Get a vector of block names for each corresponding variable. The last block is considered as the superblock and ignored.
}
\examples{
rgcca.res = list(a = rep(NA, 4))
names(rgcca.res$a) = LETTERS[1:4]
getBlocsVariables(rgcca.res)
# a, b, c
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
