% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{getFileName}
\alias{getFileName}
\title{File name from a path}
\usage{
getFileName(fi)
}
\arguments{
\item{fi}{A character giving the path of a file}
}
\value{
A character
}
\description{
Get the file name from a path
}
\examples{
fi = "/name.lastname/dirPath/fileName.tsv"
getFileName(fi)
# fileName
}
