% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getRankedValues}
\alias{getRankedValues}
\title{Rank values of a dataframe in decreasing order}
\usage{
getRankedValues(df, comp = 1, allCol = T)
}
\arguments{
\item{df}{A dataframeZ}

\item{comp}{An integer giving the index of the analysis components}

\item{allCol}{A boolean to use all the column of the datafram}
}
\description{
Rank values of a dataframe in decreasing order
}
