% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mean.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute NA}
\usage{
impute_mean(df)
}
\arguments{
\item{df}{A matrix containing non availables data}
}
\value{
A matrix with imputed values
}
\description{
Impute non available by means
}
\examples{
df = cbind(runif(9), runif(9))
df = rbind(df, c(NA, NA))
impute_mean(df)
}
