% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeEM.R
\name{imputeEM}
\alias{imputeEM}
\title{imputeEM: impute with superblock method}
\usage{
imputeEM(A, C, tau, ni = 50, tol = 1e-08, graph = FALSE,
  ncomp = NULL, naxis = 1, scale = TRUE, sameBlockWeight = TRUE,
  scheme = "centroid", bias = TRUE, superblock = FALSE,
  verbose = FALSE, tolEM = 0.001, reg = "y")
}
\arguments{
\item{A}{A matrix with J dimensions (to be changed for superblock ?)}

\item{C}{A matrix of connection}

\item{tau}{A vector of tau values with the same length as A}

\item{ni}{An integer for the maximal number of iterations before convergence}

\item{tol}{The stopping value for convergence.}

\item{graph}{if graph = TRUE, graphics are plotted/saved}

\item{ncomp}{vector containing the number of components per block in RGCCA}

\item{naxis}{number of component to select for the estimation of missing data}

\item{scale}{If scale = TRUE, each block is standardized to zero means and unit variances (default: TRUE).}

\item{sameBlockWeight}{A logical value indicating if the different blocks should have the same weight in the analysis (default, sameBlockWeight=TRUE)}

\item{scheme}{scheme chosene for RGCCA (is not useful when superblock=TRUE)
@param bias A logical value indicating if variance should be biased or not}

\item{bias}{FALSE by default. If TRUE, estimation of variance/covariance parameters with division by n-1 instead of n}

\item{superblock}{Boolean, if TRUE, the missing values are estimated with the superblock}

\item{verbose}{FALSE by default. If TRUE, displays results of convergence for each iteration}

\item{tolEM}{threshold to be reached to assess convergence.}

\item{reg}{'y' by default. Reconstruction is made by regression of the A rows on w ('w'),the A columns on Y ('y')  or without any regression with yw ('no')}
}
\value{
\item{A}{A list of blocks imputed}

\item{stab}{Convergence criterion : abs(1-obj_k/obj_{k-1})}

\item{obj}{Vector containing the mean square error between the predict values and the original non missing values at each iteration}

\item{crit}{RGCCA criterion}

\item{moy}{Estimated mean obtained for each variable  (required for the addNoise function)}

\item{stdev}{Estimated standard deviations of each variable obtained  (required for the addNoise function)}

\item{sigma}{Estimated standard deviations for the noise obtained  (required for the addNoise function)}

\item{indNA}{Position of missing values}
}
\description{
Impute with superblock method
}
\details{
This method is used for the implementation of EM algorithm for missing data
}
