% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSB.R
\name{imputeSB}
\alias{imputeSB}
\title{imputeSB: impute with superblock method}
\usage{
imputeSB(
  A,
  tau = rep(1, length(A)),
  ni = 50,
  tol = 1e-08,
  graph = FALSE,
  ncomp = NULL,
  naxis = 1,
  scale = TRUE,
  sameBlockWeight = TRUE,
  bias = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{A}{A list of J blocks}

\item{tau}{A vector of tau values with the same length as A}

\item{ni}{Number of iterations}

\item{tol}{The stopping value for convergence.}

\item{graph}{if graph = TRUE,}

\item{ncomp}{vector containing the number of components per block in RGCCA}

\item{naxis}{number of component to select in the superblock for the estimation of missing data}

\item{scale}{If scale = TRUE, each block is standardized to zero means and unit variances (default: TRUE).}

\item{sameBlockWeight}{A logical value indicating if the different blocks should have the same weight in the analysis (default, sameBlockWeight=TRUE)}

\item{bias}{A logical value indicating if variance should be biased or not}

\item{verbose}{If TRUE, displays the differents RGCCA steps}
}
\value{
\item{A}{A list of blocks imputed}

\item{crit}{Convergence criterion : abs(1-obj_k/obj_{k-1})}

\item{obj}{Vector containing the mean square error between the predict values and the original non missing values at each iteration}
}
\description{
Impute with superblock method
}
\details{
This method is used for the implementation of EM algorithm for missing data
}
