% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.character2.r
\name{is.character2}
\alias{is.character2}
\title{Test for character vector}
\usage{
is.character2(x, type = "any")
}
\arguments{
\item{x}{A matrix or a vector}
}
\value{
A bolean for the presence (FALSE) or the absence (TRUE) of at least
one quantitative variable
}
\description{
Tests if a dataframe is composed only by qualitative variables
}
\examples{
x = matrix(c(runif(10), LETTERS[seq(10)]), 10, 2)
is.character2(x)
# TRUE
is.character2(LETTERS[seq(10)])
# TRUE
}
