% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{isCharacter}
\alias{isCharacter}
\title{Test for character vector}
\usage{
isCharacter(x)
}
\arguments{
\item{x}{A matrix or a vector}
}
\value{
A bolean for the presence (FALSE) or the absence (TRUE) of at least one quantitative variable
}
\description{
Tests if a dataframe is composed only by qualitative variables
}
\examples{
x = matrix(c(runif(10), LETTERS[1:10]), 10, 2)
isCharacter(x)
# FALSE TRUE
isCharacter(LETTERS[1:10])
# TRUE
}
