% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_common_rows.R
\name{keep_common_rows}
\alias{keep_common_rows}
\title{Keep only the rows with the same names among a list of dataframe}
\usage{
keep_common_rows(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A list of dataframe
}
\description{
Keep only the rows with the same names among a list of dataframe
}
