% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{keepCommonRow}
\alias{keepCommonRow}
\title{Keep the common row names}
\usage{
keepCommonRow(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A list of datrame with the same number of row (those with the same name)
}
\description{
Keep the common row names
}
