% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_blocks.R
\name{load_blocks}
\alias{load_blocks}
\title{Create a list of matrix from loading files corresponding to blocks}
\usage{
load_blocks(file, names = NULL, sep = "\\t", header = TRUE,
  rownames = 1)
}
\arguments{
\item{file}{A character giving the path of a file}

\item{names}{A character giving a list of names for the blocks}

\item{sep}{A character giving the column separator}

\item{header}{A bolean giving the presence or the absence of the header}

\item{rownames}{An integer corresponding to the column number of the row
names (NULL otherwise)}
}
\value{
A list matrix corresponding to the blocks
}
\description{
Create a list of matrix from loading files corresponding to blocks
}
\examples{
\dontrun{
load_blocks (TRUE,
    "inst/extdata/agriculture.tsv,inst/extdata/industry.tsv,inst/extdata/politic.tsv",
    "agric,ind,polit")
}
}
