% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_response.R
\name{load_response}
\alias{load_response}
\title{Create a matrix corresponding to the response}
\usage{
load_response(blocks = NULL, file = NULL, separator = "\\t",
  header = TRUE, rownames = 1, decimal = ".")
}
\arguments{
\item{blocks}{A list of matrix}

\item{file}{A character giving the path of a file}

\item{separator}{A character giving the column separator}

\item{header}{A bolean giving the presence or the absence of the header}

\item{rownames}{An integer corresponding to the column number of the row#'}

\item{decimal}{A character giving the decimal separator}
}
\value{
A matrix corresponding to the response
}
\description{
Create a matrix corresponding to the response
}
\examples{
\dontrun{
blocks = lapply(seq(3), function(x) matrix(runif(47 * 5), 47, 5))
load_response (blocks, 'inst/extdata/response3.tsv')
}
}
