% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{loadData}
\alias{loadData}
\title{Creates a matrix from loading a file}
\usage{
loadData(f, sep = "\\t", row.names = 1, h = TRUE)
}
\arguments{
\item{f}{A character giving the file name}

\item{sep}{A character giving the column separator}

\item{row.names}{A vector of characters giving the names of the rows}

\item{h}{A bolean giving the presence or the absence of the header}
}
\value{
A matrix containing the loaded file
}
\description{
Creates a matrix from loading a file
}
\examples{
\dontrun{
loadData("data/agriculture.tsv")
}
}
