% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rgcca.R
\name{model_rgcca}
\alias{model_rgcca}
\title{Model RGCCA}
\usage{
model_rgcca(fit, labels = labels)
}
\arguments{
\item{fit}{A fitted rgccac object}

\item{labels}{boolean , if labels = TRUE, add the correlation
between latents and manifests variables 
and the betas/gammas between the latents nodes}
}
\value{
a string which will be used as a basis for the creation of the graph
}
\description{
Model RGCCA
}
\examples{
model_rgcca(fit,labels = TRUE)
}
