% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_df.R
\name{order_df}
\alias{order_df}
\title{Rank values of a dataframe in decreasing order}
\usage{
order_df(df, comp = 1, allCol = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{comp}{An integer giving the index of the analysis components}

\item{allCol}{A boolean to use all the column of the dataframe}
}
\value{
A datafram with ordered values
}
\description{
Rank values of a dataframe in decreasing order
}
\examples{
df = sapply(seq(2), function(x) runif(10))
order_df(df)
}
