% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerplot.R
\name{outerplot}
\alias{outerplot}
\title{Dynamic Outer Network Visualization for fitted RGCCAC object}
\usage{
outerplot(
  fit,
  color.background = c("white", "white"),
  color.border = c("black", "black"),
  shape = c("oval", "box"),
  color.edge = c("black", "black"),
  display_numbers = TRUE,
  configure = FALSE,
  num = num
)
}
\arguments{
\item{fit}{a fitted rgccac object}

\item{color.background}{string's vector of the color of the node's 
background of the latents and manifests variables.
Default to c("white", "white")}

\item{color.border}{string's vector of the color of the 
node's border of the latents and manifests variables.
Default to c("black", "black")}

\item{shape}{string's vector of the shape of the nodes
latents variables and manifests variables.
Default to c("oval", "box")}

\item{color.edge}{string's vector of the color of the 
edges between LV and MV and between LV.
Default to c("black", "black")}

\item{display_numbers}{boolean if display_numbers = TRUE, 
display the cor_y_eta between latents and manifests nodes 
and the betas/gammas between the latents nodes}

\item{configure}{boolean If true, 
open a control panel that allows you to modify some parameters.}

\item{num}{string or integer. Display the outer graph of the num block or
all the blocks if num  = "all"}
}
\value{

}
\description{
Dynamic Outer Network Visualization for fitted RGCCAC object
}
