% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelize.R
\name{parallelize}
\alias{parallelize}
\title{Set a list of ocket for parralel package in Windows
f : a function to parralelize
nperm : a vector object for a lapply type function
varlist : character vector of names of objects to export}
\usage{
parallelize(varlist = c(), nperm, f, n_cores = NULL,
  envir = environment(), applyFunc = "parSapply")
}
\description{
Set a list of ocket for parralel package in Windows
f : a function to parralelize
nperm : a vector object for a lapply type function
varlist : character vector of names of objects to export
}
