% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parseList}
\alias{parseList}
\title{Convert a character in a vector}
\usage{
parseList(s)
}
\arguments{
\item{s}{A character separated by comma}
}
\value{
A vector of characters whitout spaces
}
\description{
Convert a character in a vector
}
\examples{
s = "1,2, 3"
parseList(s)
}
