% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_diagram_outer.R
\name{path_diagram_outer}
\alias{path_diagram_outer}
\title{Outer Static Visualization Network RGCCAC}
\usage{
path_diagram_outer(
  model,
  display_numbers = TRUE,
  graph_options = list(overlap = F, layout = "dot", rankdir = "UD", concentrate = T),
  edge_options = list(color = "black", arrowhead = "vee"),
  node_options_LV = list(shape = "oval", color = "black"),
  node_options_MV = list(shape = "box", color = "black"),
  num = num
)
}
\arguments{
\item{model}{fitted rgccac object}

\item{graph_options}{string,default to "dot". The layout among "dot and "fdp". 
The rankdir sets direction of graph layout. Default to "LR" 
Concentrate allows to merges multiedges into a single edge.
This feature is not  available outside of "dot" layout
 and display_number = TRUE. Default to TRUE}

\item{node_options_LV}{string. The shape and the color of the latents nodes
Default to list(shape = "oval", color = "black")}

\item{num}{string or integer. Display the outer graph of the num's block or
all the blocks if num  = "all"}

\item{isplay_numbers}{boolean, default to TRUE, if display_numbers = TRUE, 
display the cor_eta_y between latents and manifests nodes 
and the betas/gammas between the latents nodes}

\item{dge_options}{string. The color of the edges and shape of the arrow
Default to list(color = "black", arrowhead = "vee")}
}
\value{

}
\description{
path_diagram allows to display an outer graphic of an RGCCAC object
}
\examples{
path_diagram_outer (fit.rgccac, display_numbers = TRUE,
graph_options = list(overlap = F, layout = "dot", 
                    rankdir = "UD", concentrate = T),
edge_options = list(color = "black", arrowhead = "vee"),
node_options_LV = list(shape = "oval",  color = "black"),
node_options_MV = list(shape = "box",  color = "black"),num = "all")
}
