% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cval.r
\name{plot.cval}
\alias{plot.cval}
\title{plot.cval}
\usage{
\method{plot}{cval}(x, bars = "sd", alpha = 0.05, ...)
}
\arguments{
\item{x}{result of rgcca_cv function (see \link{rgcca_cv})}

\item{bars}{among "sd" for standard deviations, "stderr" for standard error (standard deviations divided by sqrt(n), ci for confidence interal, cim for confidence interval of the mean.}

\item{alpha}{used for confidence interval bars (ci or cim), risk. Default to 0.05}

\item{...}{Further plot options}
}
\description{
plot.cval
}
\examples{
data("Russett")
blocks <- list(
    agriculture = Russett[, seq(3)],
    industry = Russett[, 4:5],
    politic = Russett[, 6:11])
    res=rgcca_cv(blocks, type="rgcca",par="tau",par_value=c(0,0.2,0.3),n_cv=1,n_cores=1)
   plot(res)
}
