% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.list_rgcca.R
\name{plot.list_rgcca}
\alias{plot.list_rgcca}
\title{Regularized Generalized Canonical Correlation Analysis (RGCCA)}
\usage{
\method{plot}{list_rgcca}(x, type = "both", resp = rep(1,
  NROW(x$Y[[1]])), i_block = 1, i_block_y = i_block, compx = 1,
  compy = 2, remove_var = FALSE, text_var = TRUE, text_ind = TRUE,
  response_name = "Response", no_overlap = FALSE, title = NULL,
  title_var = "Variable correlations with", title_ind = "Sample space",
  n_mark = 100, collapse = FALSE, cex = 1, cex_sub = 10,
  cex_main = 14, cex_lab = 12, colors = NULL, errorbar = "CIscore",
  ...)
}
\arguments{
\item{x}{Result of rgcca function  (see\code{\link[RGCCA]{rgcca}} )}

\item{type}{Type among c("ind","var","both","ave","cor","weight","network"). "ind" for individual graph, "var" for variable graph, "both" for both, "ave" for the variance average in each block, "net"for network, "weight" for plotting the top weights in compx, or "cor" for the top correlation with the compx.}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{i_block}{An integer giving the index of a list of blocks}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{text_var}{A bolean to represent the variables with their row names (TRUE)
or with circles (FALSE)}

\item{text_ind}{A bolean to represent the individuals with their row names (TRUE)
or with circles (FALSE)}

\item{response_name}{A character giving the legend title}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{title_var}{Character for the title of the variable space}

\item{title_ind}{Character for the title of the individual space}

\item{n_mark}{An integer giving the number of top variables to select}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_main}{An integer for the size of the title}

\item{cex_lab}{An integer for the size of the axis titles}

\item{colors}{representing a vector of the colors used in the graph. Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{errorbar}{("CImean","CIscores","sd")}

\item{...}{Further graphical parameters applied to both (individual and variable) spaces}
}
\description{
plot 
Plots
}
\examples{
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric, X_ind, X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
resRgcca=rgcca(blocks=A,connection=C,
tau=rep(1,3),ncomp=rep(2,3),superblock=FALSE)
plot(resRgcca,type="cor")
plot(resRgcca,type="cor")
}
