% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca_permutation.R
\name{plot.permutation}
\alias{plot.permutation}
\title{plot.permutation}
\usage{
\method{plot}{permutation}(x, type = "zstat", cex = 1, main = NULL,
  cex_main = 25 * cex, cex_sub = 16 * cex, cex_point = 3 * cex,
  cex_lab = 19 * cex, ...)
}
\arguments{
\item{x}{result of rgcca_permutation (see  \code{\link[RGCCA]{rgcca_permutation}} )}

\item{type}{An string giving the type of the index to look at (among 'crit' for
the RGCCA criterion and 'zstat' for the pseudo Z-score)}

\item{cex}{general size of the text}

\item{main}{title of the plot}

\item{cex_main}{= 25 * cex, size of the main text (title)}

\item{cex_sub}{= 16 * cex, size of the subtitle text}

\item{cex_point}{= 3 * cex, size of the point}

\item{cex_lab}{= 19 * cex, size of the labels}

\item{...}{Further graphical parameters}
}
\description{
plot.permutation
}
\examples{
data("Russett")
A = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
perm <- rgcca_permutation(A, nperm = 2, n_cores = 1)
plot(perm)
perm <- rgcca_permutation(A, perm.par = "sparsity", nperm = 2, n_cores = 1)
plot(perm)
}
