% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca.R
\name{plot.rgcca}
\alias{plot.rgcca}
\title{Regularized Generalized Canonical Correlation Analysis (RGCCA)}
\usage{
\method{plot}{rgcca}(x, resp = rep(1, NROW(x$Y[[1]])), i_block = 1,
  compx = 1, compy = 2, remove_var = FALSE, ...)
}
\arguments{
\item{x}{Result of rgcca function}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{i_block}{An integer giving the index of a list of blocks}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{...}{parameters from plot_ind or plot_var_2D}
}
\description{
plot 
Plots
}
\examples{
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric, X_ind, X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
resRgcca=rgcca(blocks=A,connection=C,
tau=rep(1,3),ncomp=rep(2,3),superblock=FALSE)
plot(resRgcca)
}
