% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWhichNAmethod.r
\name{plot.whichNAmethod}
\alias{plot.whichNAmethod}
\title{Plots the impact of increasing missing data on RGCCA}
\usage{
\method{plot}{whichNAmethod}(x, output = "rv", ylim = NULL,
  block = "all", barType = "sd", main = NULL, ylab = "", ...)
}
\arguments{
\item{x}{A list resulting of which NA method}

\item{output}{="rv": Can be also "a" for correlations between axes, "bm" for the percent of similar biomarkers, "rvComplete" if the RV is calculated only on complete dataset, or "rmse" for Root Mean Squares Error.}

\item{ylim}{=c(0.8,1) y limits}

\item{block}{="all" or a number indicating the position of the chosen block in the initial list}

\item{barType}{="sd" or "stderr". Indicates which error bar to build}

\item{main}{=NULL Name of the file}

\item{ylab}{label of y-axis}
}
\description{
Plots the impact of increasing missing data on RGCCA
}
\examples{
set.seed(42);X1=matrix(rnorm(350),70,5);X2=matrix(rnorm(280),70,4);X1[1,1]=NA;X2[2,]=NA
colnames(X1)=paste("A",1:5);colnames(X2)=paste("B",1:4); 
rownames(X1)=rownames(X2)=paste0("S",1:70);A=list(X1,X2);
listResults=whichNAmethod(blocks=A,patternNA=c(0.1,0.2),
listMethods=c("mean","complete","nipals","knn4"))
plot(x=listResults,ylim=c(0,1),output="a")
}
