% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Plot of components space}
\usage{
plot2D(rgcca, df, title = "Biplot", group, name_group = "Response",
  compx = 1, compy = 2, i_block = NULL, p = NULL, text = TRUE,
  i_block_y = i_block, colours = c("blue", "gray", "#cd5b45"),
  collapse = FALSE, no_overlap = FALSE, cex = 1, cex_sub = 16 *
  cex, cex_point = 3 * cex, cex_lab = 19 * cex)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{group}{A vector of character with levels used to color the points}

\item{name_group}{A character giving the type of groups (either "Blocs" or
"Response")}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{p}{A ggplot object}

\item{text}{A bolean to represent the points with their row names (TRUE)
or with circles (FALSE)}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{colours}{A vectof of character to color quantitative dat}

\item{collapse}{A boolean to combine the variables of each blocks as result}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Plots RGCCA components in a bi-dimensional space
}
\examples{
#df = as.data.frame(matrix(runif(20*2, min = -1), 20, 2))
#AVE = lapply(seq(4), function(x) runif(2))
#rgcca_out = list(AVE = list(AVE_X = AVE))
#plot2D(rgcca_out, df, "Samples", rep(c("a","b"), each=10), "Response")
data(Russett)
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
politic = Russett[, 6:11] )
# rgcca_out = rgcca(blocks)
# plot2D(rgcca_out)
}
