% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{Plot a bootstrap}
\usage{
plot_bootstrap(df, rgcca, show_boot = TRUE, n_mark = 30,
  colors = NULL, ...)
}
\arguments{
\item{df}{A dataframe}

\item{rgcca}{A list giving the results of a R/SGCCA}

\item{show_boot}{A boolean to show the bootstrap mean and sd on the graphic}

\item{n_mark}{An integer giving the number of top variables to select}

\item{colors}{reoresenting a vector of colors}
}
\description{
Plot the top variables from a bootstrap
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, 2, FALSE, n_cores = 1)
selected.var = get_bootstrap(rgcca_out, boot, n_cores = 1)
plot_bootstrap(selected.var, rgcca_out)
}
