% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_1D.R
\name{plot_bootstrap_1D}
\alias{plot_bootstrap_1D}
\title{Plot a bootstrap in 1D}
\usage{
plot_bootstrap_1D(b, x = "br", y = "occ", n = 50, cex = 1,
  cex_sub = 16 * cex, cex_axis = 10 * cex)
}
\arguments{
\item{b}{A matrix of boostrap}

\item{x}{A character for the column used in the plot}

\item{y}{A character for the column to color the bars}

\item{n}{An integer giving the number maximum of top variables}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_axis}{An integer for the size of the axis text}
}
\description{
Histogram of the best variables of an RGCCA bootstrap with, on the x-axis,
the number of non-zero occurrences (SGCCA) or the bootstrap-ratio
(mean/sd; RCCA). The bars are colored according to the average weight of
the boostrap  (according to an ascending gradient from red to blue)
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks, tau = 0.75, type = "sgcca")
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(rgcca_out, boot, n_cores = 1)
plot_bootstrap_1D(selected.var)
rgcca_out = rgcca.analyze(blocks)
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(rgcca_out, boot, n_cores = 1)
plot_bootstrap_1D(selected.var)
}
