% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram settings}
\usage{
plot_histogram(p, df, title = "Histogram", group = NA, colors = NULL,
  cex = 1, cex_main = 25 * cex, cex_sub = 16 * cex, cex_axis = 10 *
  cex)
}
\arguments{
\item{p}{A ggplot object}

\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{group}{A vector of character giving the group for the rows}

\item{colors}{reoresenting a vector of colors}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_axis}{An integer for the size of the axis text}

\item{low_col}{A character giving the color used for the lowest part of
the gradient}

\item{high_col}{A character giving the color used for the highest part of
the gradient}

\item{mid_col}{A character giving the color used for the middle part of
the gradient}
}
\description{
Default font for a vertical barplot.
}
\examples{
df = data.frame(x = runif(30), order = 30:1)
library("ggplot2")
p = ggplot(df, aes(order, x))
plot_histogram(p, df, "This is my title")
# Add colors per levels of a variable
df$color = rep(c(1,2,3), each=10)
p = ggplot(df, aes(order, x, fill = color))
plot_histogram(p, df, "Histogram", as.character(df$color))
}
