% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot the two components of a RGCCA}
\usage{
plot_ind(
  rgcca_res,
  resp = rep(1, NROW(rgcca_res$Y[[1]])),
  compx = 1,
  compy = 2,
  i_block = length(rgcca_res$Y),
  text = TRUE,
  i_block_y = i_block,
  response_name = "Response",
  no_overlap = FALSE,
  predicted = NULL,
  title = NULL,
  legend = TRUE,
  ...
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{compx}{An integer giving the index of the block component for the x-axis}

\item{compy}{An integer giving the index of the block component for the y-axis}

\item{i_block}{An integer giving the index of the block}

\item{text}{A logical value indicating if rowanmes are displayed (default = TRUE)}

\item{i_block_y}{An integer giving the index of the block (different from i_block)}

\item{response_name}{A character giving the legend title}

\item{no_overlap}{A logical value indicating if text-overlappin is accepted}

\item{predicted}{A list containing as  2nd element a matrix of predicted components}

\item{title}{A character giving the title of the plot}

\item{legend}{A logical value indicatif if the legend should be plotted}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
Plot the two components of a RGCCA
}
\examples{
coord = lapply(seq(3),
   function(x) matrix(runif(15 * 2, min = -1), 15, 2))
AVE_X = lapply(seq(3), function(x) runif(2))
for (i in 1:length(coord))
row.names(coord[[i]]) = seq(15)
rgcca_out = list(Y = coord, AVE = list(AVE_X = AVE_X),
call = list(blocks = coord, ncomp = rep(2,3))) # TODO
# Using a superblock
resp = as.matrix(rep(LETTERS[seq(3)], each = 5))
row.names(resp) = seq(15)
rgcca_out$call$method="rgcca"
class(rgcca_out) = "rgcca"
plot_ind(rgcca_out, resp)
# Using the first block
resp = as.matrix(runif(15, min=-15, max = 15))
row.names(resp) = seq(15)
plot_ind(rgcca_out, resp, 1, 2, 1)
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")])
X_ind = as.matrix(Russett[,c("gnpr","labo")])
X_polit = as.matrix(Russett[ , c("demostab", "dictator")])
A = list(X_agric, X_ind, X_polit)
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
result.rgcca = rgcca(A, connection=C, tau = c(1, 1, 1), scheme = "factorial",
scale = TRUE,ncomp=rep(2,3))
plot_ind(result.rgcca,i_block=1)
}
