% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network2.R
\name{plot_network2}
\alias{plot_network2}
\title{Plot the connection between blocks (dynamic plot)}
\usage{
plot_network2(rgcca)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}
}
\value{
A dataframe with tuples of connected blocks
}
\description{
Plot the connection between blocks (dynamic plot)
}
\examples{
library(visNetwork)
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks)
plot_network2(rgcca_out)
}
