% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permut_3D.R
\name{plot_permut_3D}
\alias{plot_permut_3D}
\title{Plot permuation in 3D}
\usage{
plot_permut_3D(perm, type = "zstat", sign = FALSE, i_block = 1,
  i_block_y = 2, i_block_z = 3, cex = 1, cex_point = 3 * cex,
  cex_lab = 19 * cex)
}
\arguments{
\item{perm}{A permutation object from a RGCCA analyse}

\item{type}{A character for the type of plot : either "ind" for individual plot or "var" for corcircle}

\item{sign}{A boolean to color by groups of alpha = 0.05, 0.01 or 0.001}

\item{i_block}{An integer giving the index of a list of blocks}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{i_block_z}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Plot permuation in 3D
}
