% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_1D.R
\name{plot_var_1D}
\alias{plot_var_1D}
\title{Barplot of a fingerprint}
\usage{
plot_var_1D(
  rgcca_res,
  comp = 1,
  n_mark = 30,
  i_block = length(rgcca_res$a),
  type = "cor",
  collapse = FALSE,
  title = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{comp}{An integer giving the index of the analysis components}

\item{n_mark}{The number of top variables to display in the correlation
circle}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A character giving the criterion to selects variables : either
"cor" for correlation between the component and the block
or "weight" for the weight of the RGCCA}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{title}{A character giving the title of the plot}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{...}{Further graphical parameters (see plot2D)}
}
\description{
Barplot of the higher outer weight vectors for a component of a block
(by default, the superblock or the last one) analysed by R/SGCCA
}
\examples{
weights = lapply(seq(3), function(x) matrix(runif(7*2), 7, 2))
for (i in seq(3))
row.names(weights[[i]]) <- paste0(letters[i],
     letters[seq(NROW(weights[[i]]))])
weights[[4]] = Reduce(rbind, weights)
rgcca_out = list(a = weights, call = list(type="rgcca", ncomp = rep(2,4)))
names(rgcca_out$a) = LETTERS[seq(4)]
rgcca_out$call$blocks = lapply(rgcca_out$a, t)
rgcca_out$call$superblock = TRUE
class(rgcca_out) = "rgcca"
# With the 1rst component of the superblock
plot_var_1D(rgcca_out, 1, type = "weight")
# With the 2nd component of the 1rst block by selecting the ten higher weights
plot_var_1D(rgcca_out, 2, 10, 1, type = "weight")
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
plot_var_1D(rgcca_out, collapse = TRUE)
}
\seealso{
\code{\link[RGCCA]{rgccad}}, \code{\link[RGCCA]{sgcca}}
}
