% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_2D.R
\name{plot_var_2D}
\alias{plot_var_2D}
\title{Plot of variables space}
\usage{
plot_var_2D(
  rgcca_res,
  compx = 1,
  compy = 2,
  i_block = length(rgcca_res$a),
  text = TRUE,
  remove_var = TRUE,
  n_mark = 30,
  collapse = FALSE,
  no_overlap = FALSE,
  title = NULL,
  resp = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{compx}{An integer giving the index of the block component for the x-axis}

\item{compy}{An integer giving the index of the block component for the y-axis}

\item{i_block}{An integer giving the index of the block}

\item{text}{A logical value indicating if rowanmes are displayed (default = TRUE)}

\item{remove_var}{logical value. If TRUE, the top 100 variables with the
highest with the block components are displayed}

\item{n_mark}{The number of top variables to display in the correlation
circle}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{no_overlap}{A logical value indicating if text-overlappin is accepted}

\item{title}{A character giving the title of the plot}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{colors}{Either a vector of integers (each integer corresponding to a
color) or of characters corresponding to names of colors (as "blue",see
colors()) or RGB code ("#FFFFFF").}

\item{...}{Further graphical parameters (see plot2D)}
}
\description{
Correlation circle highlighting the contribution of each variables to the
construction of the RGCCA components
}
\examples{
setMatrix = function(nrow, ncol, iter = 3)
 lapply(
     seq(iter),
     function(x) {
         y <- matrix(runif(nrow * ncol), nrow, ncol)
         rownames(y) <- paste0("S",1:nrow)
         return(y)
     })
blocks = setMatrix(10, 5)
blocks[[4]] = Reduce(cbind, blocks)
for (i in seq(4)) {
 colnames(blocks[[i]]) = paste0( LETTERS[i],
 as.character(seq(NCOL(blocks[[i]]))))
}
coord = setMatrix(10, 2, 4)
a = setMatrix(5, 2)
a[[4]] = setMatrix(15, 2, 1)[[1]]
AVE_X = lapply(seq(4), function(x) runif(2))
rgcca_out = list(Y = coord, a = a, AVE = list(AVE_X = AVE_X),
                 call = list(blocks = blocks))
names(rgcca_out$a) <- LETTERS[seq(4)] -> names(rgcca_out$call$blocks)
rgcca_out$call$method="rgcca"
# Using a superblock
rgcca_out$call$superblock = TRUE
rgcca_out$call$ncomp = rep(3, 4)
class(rgcca_out) = "rgcca"
plot_var_2D(rgcca_out, 1, 2)
# Using the first block
plot_var_2D(rgcca_out, 1, 2, 1)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)],
              industry = Russett[, 4:5],
              politic = Russett[, 6:11] )

rgcca_out = rgcca(blocks,ncomp=2)

# Without superblock but with the of all variables to the first block
plot_var_2D(rgcca_out, collapse = TRUE)

}
