% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAVE}
\alias{plotAVE}
\title{Plot Average Variance Explained}
\usage{
plotAVE(rgcca, i_comp)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{i_comp}{An integer giving the index of the analysis component}
}
\description{
Histogram of the Average Variance Explained for each blocks and ordered deacreasingly
}
\examples{
AVE = lapply(1:4, function(x) runif(1))
rgcca.res$AVE$AVE_X = AVE
plotAVE(rgcca.res, 1)
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
