% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.type.R
\name{plotBootstrap}
\alias{plotBootstrap}
\title{list of list weights (one per bootstrap per blocks)}
\usage{
plotBootstrap(W, comp = 1, n_mark = 100, i_block = NULL)
}
\arguments{
\item{comp}{An integer giving the index of the analysis components}
}
\description{
list of list weights (one per bootstrap per blocks)
}
