% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot Histograms}
\usage{
plotHistogram(p, df, title, color)
}
\arguments{
\item{p}{A ggplot object.}

\item{df}{A dataframe with a column named "order"}

\item{title}{A character string giving a graphic title.}

\item{color}{A vector of character giving the colors for the rows.}
}
\description{
Default font for a vertical barplot.
}
\seealso{
\code{\link[RGCCA]{rgcca}}
}
