% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpace}
\alias{plotSpace}
\title{Plot bi-dimensional space}
\usage{
plotSpace(rgcca, df, title, group, name_group, compX, compY, i_block,
  p = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "variables" or "samples")}

\item{group}{A vector color the points with a vector}

\item{name_group}{type of groups (Blocs/Response)}

\item{compX}{component used for the x-axis}

\item{compY}{component used for the y-axis}

\item{i_block}{index of the block}
}
\description{
Projects coordinates of points in a bi-dimensional space
}
