% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVariablesSpace}
\alias{plotVariablesSpace}
\title{Plot of variables space}
\usage{
plotVariablesSpace(rgcca, blocks, compX = 1, compY = 2,
  superblock = NULL, i_block = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{blocks}{A list of matrix}

\item{compX}{An integer giving the index of the analysis component used for the x-axis}

\item{compY}{An integer giving the index of the analysis component used for the y-axis}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of a superblock}

\item{i_block}{An integer giving the index of a list of blocks}
}
\description{
Correlation circle highlighting the contribution of each variables to the construction of the RGCCA components
}
\examples{
setMatrix = function(nrow, ncol, iter = 3) lapply(1:iter, function(x) matrix(runif(nrow * ncol), nrow, ncol))
blocks = setMatrix(10, 5)
blocks[[4]] = Reduce(cbind, blocks)
for (i in 1:4)
    colnames(blocks[[i]]) = paste( LETTERS[i], as.character(1:NCOL(blocks[[i]])), sep="" )
coord = setMatrix(10, 2, 4)
a = setMatrix(5, 2)
a[[4]] = matrix(runif(15 * 2), 15, 2)
AVE_X = lapply(1:4, function(x) runif(2))
rgcca.res = list(Y = coord, a = a, AVE = list(AVE_X = AVE_X))
names(rgcca.res$a) = LETTERS[1:4]
# Using a superblock
plotVariablesSpace(rgcca.res, blocks, 1, 2, TRUE)
# Using the first block
plotVariablesSpace(rgcca.res, blocks, 1, 2, FALSE, 1)
}
