% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootstrap.r
\name{print.bootstrap}
\alias{print.bootstrap}
\title{Print rgcca_bootstrap}
\usage{
\method{print}{bootstrap}(x, type = "weights", empirical = TRUE, display_order = FALSE, ...)
}
\arguments{
\item{x}{A fitted rgcca_bootstrap object
(see \code{\link[RGCCA]{rgcca_bootstrap}})}

\item{type}{Character string indicating the bootstrapped object to print:
block-weight vectors ("weights", default) or block-loading vectors
("loadings").}

\item{empirical}{A logical value indicating if the bootstrap confidence
intervals and p-values are derived from the empirical distribution.
(defaut: TRUE)}

\item{display_order}{A logical value for ordering the variables}

\item{...}{Further arguments in print
the means, 95\% intervals, bootstrap ratio, p-values and other statistics.}
}
\value{
none
}
\description{
Print a rgcca_bootstrap object
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
fit.rgcca <- rgcca(blocks, ncomp = c(2, 1, 2))
boot.out <- rgcca_bootstrap(fit.rgcca, n_boot = 20, n_cores = 2)
print(boot.out)
}
