% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rgcca.R
\name{print.rgcca}
\alias{print.rgcca}
\title{print.rgcca
Print rgcca results}
\usage{
\method{print}{rgcca}(x, ...)
}
\arguments{
\item{x}{a result of rgcca function}

\item{...}{other parameters}
}
\description{
print.rgcca
Print rgcca results
}
\examples{
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric, X_ind, X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
res = rgcca(A, C, ncomp=rep(2,3),tau = c(1, 1, 1), scheme = "factorial", scale = TRUE,verbose=FALSE)
print(res)
}
