% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{printAxis}
\alias{printAxis}
\title{Print the variance of a component}
\usage{
printAxis(rgcca, n = NULL, i = NULL, outer = FALSE)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{n}{An integer giving the index of the analysis component}

\item{i}{An integer giving the index of a list of blocks}
}
\description{
Prints the percent of explained variance for a component of a block (by default, the superblock or the last one) analysed by R/SGCCA
}
\examples{
AVE = list(c(0.6, 0.5), c(0.7, 0.45))
rgcca.res = list(AVE = list(AVE_X = AVE))
# For the superblock (or the last block)
printAxis(rgcca.res, 1)
# "Axis 1 (70\%)"
# For the first block
printAxis(rgcca.res, 2, 1)
# "Axis 2 (50\%)"
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
