% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_l1_l2.R
\name{proj_l1_l2}
\alias{proj_l1_l2}
\title{Projection onto the intersection of the L1 and L2 balls}
\usage{
proj_l1_l2(argu, a = 1)
}
\arguments{
\item{argu}{numeric vector to project onto the intersection of the L1 and
L2 balls}

\item{a}{numeric scalar defining the sparsity constraint}
}
\value{
\item{k}{number of iteration.}

\item{lambda}{threshold value for the soft-thresholding operator.}

\item{l2_SATlambda}{Boolean value indicating whether the
l2-constraint are saturated or not.}
}
\description{
Projection onto the intersection of the L1 and L2 balls
}
