% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_null_sd.R
\name{remove_null_sd}
\alias{remove_null_sd}
\title{Remove column having a standard deviation equals to 0}
\usage{
remove_null_sd(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A list of dataframe
}
\description{
Remove column having a standard deviation equals to 0
}
