% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{removeColumnSdNull}
\alias{removeColumnSdNull}
\title{Remove column having a standard deviation equals to 0}
\usage{
removeColumnSdNull(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A list of dataframe
}
\description{
Remove column having a standard deviation equals to 0
}
