% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcca_transform.R
\name{rgcca_transform}
\alias{rgcca_transform}
\title{RGCCA Transform}
\usage{
rgcca_transform(rgcca_res, blocks_test = rgcca_res$call$blocks)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}}).}

\item{blocks_test}{A list of either dataframes or matrices to be projected.}
}
\value{
A list of matrices containing the projections
of the test blocks using the canonical components from the fitted RGCCA
object.
}
\description{
Project blocks of data on canonical components extracted with a RGCCA model.
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, 1:3],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
C <- connection <- 1 - diag(3)
A <- lapply(blocks, function(x) x[1:32, ])
fit.rgcca <- rgcca(A,
  connection = C, tau = c(0.7, 0.8, 0.7),
  ncomp = c(3, 2, 4), scale_block = FALSE, superblock = FALSE
)
X <- lapply(blocks, function(x) x[39:47, ])
projection <- rgcca_transform(fit.rgcca, X)
}
