% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgccak.R
\name{rgccak}
\alias{rgccak}
\title{Internal function for computing the RGCCA parameters (RGCCA block components, outer weight vectors, etc.).}
\usage{
rgccak(A, C, tau = "optimal", scheme = "centroid", verbose = FALSE,
  init = "svd", bias = TRUE, tol = 1e-08, na.rm = TRUE,
  estimateNA = "no", scale = TRUE, sameBlockWeight = TRUE,
  initImpute = "rand")
}
\arguments{
\item{A}{A list that contains the \eqn{J} blocks of variables. Either the blocks (\eqn{X_1, X_2, ..., X_J}) or the residual matrices (\eqn{X_{h1}, X_{h2}, ..., X_{hJ}}).}

\item{C}{A design matrix that describes the relationships between blocks. (Default: complete design).}

\item{tau}{A \eqn{1 * J} vector that contains the values of the shrinkage parameters \eqn{\tau_j}, \eqn{ j=1, ..., J}. (Default: \eqn{\tau_j = 1}, \eqn{ j=1, ..., J}).
If tau = "optimal" the shrinkage intensity paramaters are estimated using the Schafer and Strimmer (2005) 
analytical formula.}

\item{scheme}{The value is "horst", "factorial", "centroid" or any diffentiable convex scheme function g designed by the user (default: "centroid").}

\item{verbose}{Will report progress while computing if verbose = TRUE (default: TRUE).}

\item{init}{The mode of initialization to use in the RGCCA algorithm. The alternatives are either by Singular Value Decompostion or random (default : "svd").}

\item{bias}{A logical value for either a biaised or unbiaised estimator of the var/cov.}

\item{tol}{Stopping value for convergence.}

\item{na.rm}{If TRUE, NIPALS algorithm taking missing values into account is run. RGCCA is run only on available data.}

\item{estimateNA}{to choose between "no","first","iterative","superblock","new","lebrusquet") TO BE DEVELOPPED}

\item{scale}{TRUE if scaled (used in superblock method)}

\item{sameBlockWeight}{useful in lebrusquet}

\item{initImpute}{'rand' or 'mean': initialization for optimization method}
}
\value{
\item{Y}{A \eqn{n * J} matrix of RGCCA outer components}

\item{Z}{A \eqn{n * J} matrix of RGCCA inner components}

\item{a}{A list of outer weight vectors}

\item{crit}{The values of the objective function to be optimized in each iteration of the iterative procedure.}

\item{AVE}{Indicators of model quality based on the Average Variance Explained (AVE): 
AVE(for one block), AVE(outer model), AVE(inner model).}

\item{call}{Call of the function}

\item{crit}{A vector that contains the values of the objective function at each iterations.}

\item{tau}{\eqn{1 * J} vector containing the value for the tau penalties applied to each of the \eqn{J} blocks of data (user specified)}
}
\description{
The function rgccak() is called by rgccad() and does not have to be used by the user.
The function rgccak() computes the RGCCA block components, outer weight vectors, etc., 
for each block and each dimension. Depending on the dimensionality of each block \eqn{X_j , j = 1, ..., J}, 
the primal (when \eqn{n > p_j}) or the dual (when \eqn{n < p_j}) algorithm is used (see Tenenhaus et al. 2015)
}
\references{
Tenenhaus M., Tenenhaus A. and Groenen PJF (2017), Regularized generalized canonical correlation analysis: A framework for sequential multiblock component methods, Psychometrika, in press

Tenenhaus A., Philippe C., & Frouin V. (2015). Kernel Generalized Canonical Correlation Analysis. Computational Statistics and Data Analysis, 90, 114-131.

Tenenhaus A. and Tenenhaus M., (2011), Regularized Generalized Canonical Correlation Analysis, Psychometrika, Vol. 76, Nr 2, pp 257-284.

Schafer J. and Strimmer K., (2005), A shrinkage approach to large-scale covariance matrix estimation and implications for functional genomics. Statist. Appl. Genet. Mol. Biol. 4:32.
}
