% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot object}
\usage{
save_plot(f, p)
}
\arguments{
\item{f}{A character giving the name of a file}

\item{p}{A ggplot object}
}
\description{
Save a ggplot in various output formats
}
\examples{
df <- as.data.frame(matrix(runif(20), 10, 2))
# p = ggplot(df, aes(df[, 1], df[, 2]))
# save_plot('Rplot.png', p)
}
