% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{savePlot}
\alias{savePlot}
\title{Save a ggplot object}
\usage{
savePlot(f, p)
}
\arguments{
\item{f}{A character giving the name of a file}

\item{p}{A ggplot object}
}
\description{
Save a ggplot in various output formats
}
\examples{
library("ggplot2")
df = as.data.frame(matrix(runif(20), 10, 2))
p = ggplot(df, aes(df[, 1], df[, 2]))
savePlot("Rplot.png", p)
}
