% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale3.R
\name{scale3}
\alias{scale3}
\title{scale3}
\usage{
scale3(A, center = TRUE, scale = TRUE, bias = TRUE)
}
\arguments{
\item{A}{A list of J blocks}

\item{center}{if TRUE, all variables are centered}

\item{scale}{if TRUE, all variables are scaled}

\item{bias}{if TRUE, the estimator of variance is SS/sqrt(n-1), if FALSE, it is SS/sqrt(n)}
}
\value{
\item{A}{The resulting list of matrices}
}
\description{
Uses cov3 to scale a list of matrix (taking into account the missing values)
}
\details{
This method is used for the implementation of EM algorithm for missing data
}
\examples{
X1=matrix(rnorm(15),5,3)
scale3(X1)
}
