% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.type.R
\name{select.type}
\alias{select.type}
\title{Translates the type string into the appropriate tau and function}
\usage{
\method{select}{type}(A = blocks, opt = NULL, C = 1 -
  diag(length(A)), tau = rep(1, length(A)), ncomp = rep(1, length(A)),
  scheme = "centroid", superblock = TRUE, type = "rgcca",
  verbose = TRUE, quiet = FALSE)
}
\description{
Translates the type string into the appropriate tau and function
}
