% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_analysis.R
\name{select_analysis}
\alias{select_analysis}
\title{Define the parameters associated with each multi-block component
method of the literature.}
\usage{
select_analysis(
  blocks,
  connection = 1 - diag(length(blocks)),
  tau = rep(1, length(blocks)),
  sparsity = rep(1, length(blocks)),
  ncomp = rep(1, length(blocks)),
  scheme = "centroid",
  superblock = TRUE,
  method = "rgcca",
  quiet = FALSE,
  response = NULL
)
}
\arguments{
\item{blocks}{A list that contains the J blocks of variables
\eqn{\mathbf{X_1}, \mathbf{X_2}, ..., \mathbf{X_J}}{X1, X2, ..., XJ}.
Block \eqn{\mathbf{X}_j}{Xj} is a matrix of dimension
\eqn{n \times p_j}{n x p_j} where n is the number of
observations and \eqn{p_j} the number of variables.}

\item{connection}{A symmetric matrix (\eqn{J \times J}{J x J}) that
describes the relationships between blocks.}

\item{tau}{Either a \eqn{1 \times J}{1 x J} vector or a
\eqn{\mathrm{max}(ncomp) \times J}{max(ncomp) x J} matrix containing
the values of the regularization parameters (default: tau = 1, for each
block and each dimension). The regularization parameters varies from 0
(maximizing the correlation) to 1 (maximizing the covariance). If
tau = "optimal" the regularization parameters are estimated for each block
and each dimension using the Schafer and Strimmer (2005) analytical formula.
If tau is a \eqn{1 \times J}{1 x J} vector, tau[j] is identical across the
dimensions of block \eqn{\mathbf{X}_j}{Xj}. If tau is a matrix, tau[k, j]
is associated with \eqn{\mathbf{X}_{jk}}{Xjk} (kth residual matrix for
block j). The regularization parameters can also be estimated using
\link{rgcca_permutation} or \link{rgcca_cv}.}

\item{sparsity}{Either a \eqn{1*J} vector or a \eqn{max(ncomp) * J} matrix
encoding the L1 constraints applied to the outer weight vectors. The amount
of sparsity varies between \eqn{1/sqrt(p_j)} and 1 (larger values of sparsity
correspond to less penalization). If sparsity is a vector, L1-penalties are
the same for all the weights corresponding to the same block but different
components:
\deqn{for all h, |a_{j,h}|_{L_1} \le c_1[j] \sqrt{p_j},}
with \eqn{p_j} the number of variables of \eqn{X_j}.
If sparsity is a matrix, each row \eqn{h} defines the constraints applied to
the weights corresponding to components \eqn{h}:
\deqn{for all h, |a_{j,h}|_{L_1} \le c_1[h,j] \sqrt{p_j}.} It can be
estimated by using \link{rgcca_permutation}.}

\item{ncomp}{Vector of length J indicating the number of block components
for each block.}

\item{scheme}{Character string or a function giving the scheme function for
covariance maximization among "horst" (the identity function), "factorial"
 (the squared values), "centroid" (the absolute values). The scheme function
 can be any continously differentiable convex function and it is possible to
 design explicitely the scheme function (e.g. function(x) x^4) as argument of
 rgcca function.  See (Tenenhaus et al, 2017) for details.}

\item{superblock}{Boolean indicating the presence of the superblock.
Default = TRUE}

\item{method}{A character string indicating the multi-block component
method to consider: rgcca, sgcca, pca, spca, pls, spls, cca,
ifa, ra, gcca, maxvar, maxvar-b, maxvar-a, mcoa,cpca-1, cpca-2,
cpca-4, hpca, maxbet-b, maxbet, maxdiff-b, maxdiff,
sabscor, ssqcor, ssqcov-1, ssqcov-2, ssqcov, sumcor,
sumcov-1, sumcov-2, sumcov, sabscov-1, sabscov-2.}

\item{quiet}{Logical value indicating if warning messages are reported.}

\item{response}{Numerical value giving the position of the response block.
When the response argument is filled the supervised mode is automatically
activated.}
}
\value{
\item{blocks}{List of blocks.}

\item{scheme}{Character string or a function giving the scheme
function used for covariance maximization.}

\item{penalty}{Vector of length J (or character string for
'optimal' setting) indicating the values of the tuning parameters.}

\item{ncomp}{Vector of length J indicating the number of block
components or each block.}

\item{connection}{Symmetric matrix (J*J) that describes the
relationships between blocks.}

\item{superblock}{Logical value indicating if superblock is
included in the analysis.}

\item{response}{Integer giving the value of the response block if
any. NULL otherwise.}

\item{par}{String that indicates if penalty refers to tau or
sparsity.}

\item{gcca}{Function used to compute the analysis. Either rgccad
or sgcca.}
}
\description{
Define the parameters associated with each multi-block component
method of the literature.
}
