% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{setBlocks}
\alias{setBlocks}
\title{Create a list of matrix from loading files corresponding to blocks}
\usage{
setBlocks(superblock, file, names = NULL, sep = "\\t", header = TRUE,
  rownames = ROW_NAMES)
}
\arguments{
\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of a superblock}

\item{file}{A character giving the path of a file used as a response}

\item{names}{A character giving a list of names for the blocks}

\item{sep}{A character giving the column separator}

\item{header}{A bolean giving the presence or the absence of the header}

\item{rownames}{An integer corresponding to the column number of the row names (NULL otherwise)}
}
\value{
A list matrix corresponding to the blocks
}
\description{
Create a list of matrix from loading files corresponding to blocks
}
\examples{
\dontrun{
setBlocks (TRUE, "data/agriculture.tsv,data/industry.tsv,data/politic.tsv", "agric,ind,polit")
}
}
