% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{setConnection}
\alias{setConnection}
\title{Create a matrix from loading a file corresponding to a connection between the blocks}
\usage{
setConnection(blocks, file = NULL, sep = "\\t")
}
\arguments{
\item{blocks}{A list of matrix}

\item{file}{A character giving the path of a file used as a response}

\item{sep}{A character giving the column separator}
}
\value{
A matrix corresponding to the response
}
\description{
Create a matrix from loading a file corresponding to a connection between the blocks
}
\examples{
\dontrun{
blocks = lapply(1:4, function(x) matrix(runif(47 * 5), 47, 5))
setConnection (blocks, "data/connection.tsv")
}
}
