% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{setResponse}
\alias{setResponse}
\title{Create a matrix from loading a file corresponding to the response}
\usage{
setResponse(blocks, file = NULL, sep = "\\t", header = TRUE)
}
\arguments{
\item{blocks}{A list of matrix}

\item{file}{A character giving the path of a file used as a response}

\item{sep}{A character giving the column separator}

\item{header}{A bolean giving the presence or the absence of the header}
}
\value{
A matrix corresponding to the response
}
\description{
Create a matrix from loading a file corresponding to the response
}
\examples{
\dontrun{
blocks = lapply(1:3, function(x) matrix(runif(47 * 5), 47, 5))
setResponse (blocks, "data/response3.tsv")
}
}
