% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgcca.permute.crit.R
\name{sgcca.permute.crit}
\alias{sgcca.permute.crit}
\title{Run through a set of constraint parameters c1s to select the best with permutation
Only one component per block for the time being}
\usage{
sgcca.permute.crit(A, c1s = NULL, nperm = 20, C = 1 -
  diag(length(A)), ncomp = rep(1, length(A)), scheme = "factorial",
  plot = FALSE, tol = .Machine$double.eps,
  n_cores = parallel::detectCores() - 1, scale = TRUE)
}
\arguments{
\item{A}{A list that contains the J blocks of variables X_1, X_2, ..., X_J}

\item{c1s}{A matrix containing sets of constraint variables, one row by set. If null, sgcca.permute takes 10 sets between min values ($1/sqrt(ncol)$) and 1}

\item{nperm}{Number of permutation tested for each set of constraint}

\item{C}{A design matrix that describes the relationships between blocks (default: complete design)}

\item{ncomp}{Number of component computed for each block}

\item{scheme}{The value is "horst", "factorial" or "centroid" (default: "centroid")}

\item{plot}{A logical, should a plot of coeffi}

\item{tol}{Tolerance}

\item{n_cores}{For linux and MacOS number of cores used for parallelisation}

\item{scale}{If TRUE, the data is scaled}
}
\value{
A list containing :

\item{pval}{Pvalue}

\item{zstat}{Statistic Z}

\item{bestpenalties}{Penalties corresponding to the best Z-statistic}

\item{permcrit}{RGCCA criteria obtained with permutation set}

\item{crit}{ RGCCA criterion for the original dataset}
}
\description{
Run through a set of constraint parameters c1s to select the best with permutation
Only one component per block for the time being
}
