% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{the function soft() encodes the soft-thresholding operator}
\usage{
soft(x, d)
}
\arguments{
\item{x}{Float}

\item{d}{Float}
}
\description{
the function soft() encodes the soft-thresholding operator
}
\keyword{internal}
