% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeList.R
\name{writeList}
\alias{writeList}
\title{write the different blocks of a list in a working directory}
\usage{
writeList(A, wd = getwd())
}
\arguments{
\item{A}{A list of blocks}

\item{wd}{Adress of the working directory where the files will be created}
}
\description{
The function writeList write
}
\examples{
set.seed(42);X1=matrix(rnorm(35),7,5);X2=matrix(rnorm(28),7,4)
A=list(X1,X2)
writeList(A,wd=getwd())
}
