\name{BinarySearch}
\alias{BinarySearch}
\title{Internal function which does not have to be used by the users}
\usage{
  BinarySearch(argu, sumabs)
}
\description{
  Internal function which does not have to be used by the
  users
}
\author{
  Daniela Witten, Robert Tibshirani, Sam Gross and
  Balasubramanian Narasimhan
}
\keyword{internal}

