% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Compute bootstrap}
\usage{
bootstrap(rgcca, n_boot = 5, n_cores = parallel::detectCores() - 1,
  ...)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{n_boot}{A integer for the number of boostrap}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{...}{other RGCCA parameters # TODO}
}
\value{
A list of RGCCA bootstrap weights
}
\description{
Computing boostrap of RGCCA
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
bootstrap(rgcca_out, n_boot = 2, n_cores = 1)
bootstrap(rgcca_out, n_boot = 2, n_cores = 1, blocks = lapply(blocks, scale),
 superblock = FALSE)
}
