% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Compute bootstrap}
\usage{
bootstrap(rgcca_res, n_boot = 5, n_cores = parallel::detectCores() - 1,
  ...)
}
\arguments{
\item{rgcca_res}{Result of a RGCCA (see  \code{\link[RGCCA]{rgcca}} )}

\item{n_boot}{A integer for the number of boostraps}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{...}{other RGCCA parameters # TODO}
}
\value{
A list containing two elements: bootstrap and rgcca. bootstrap is a list of produced rgccas while rgcca is the original rgcca.
}
\description{
Computing boostrap of RGCCA in order to visualize the stability of the weights found in RGCCA
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
b=bootstrap(rgcca_out, n_boot = 2, n_cores = 1)
plot(b,n_cores=1)
plot(b,type="2D",n_cores=1)
rgcca_out = rgcca(blocks, superblock = FALSE)
bootstrap(rgcca_out, n_boot = 2, n_cores = 1, blocks = lapply(blocks, scale),
 superblock = FALSE)
}
\seealso{
\code{\link[RGCCA]{plot.bootstrap}} , \code{\link[RGCCA]{print.bootstrap}}
}
