% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_k.R
\name{bootstrap_k}
\alias{bootstrap_k}
\title{Compute bootstrap (internal).}
\usage{
bootstrap_k(rgcca_res, inds = NULL)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{inds}{A vector of integers defining the index of the observations
taken into account for this bootstrap sample.}
}
\value{
\item{W}{A list of RGCCA bootstrap weights. Returned only if there
are no missing variables, otherwise the name of the missing variables are
returned.}

\item{L}{A list of RGCCA bootstrap loadings.  Returned only if there
are no missing variables, otherwise the name of the missing variables are
returned.}
}
\description{
Internal function for computing bootstrap of RGCCA.
}
\keyword{internal}
