% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_k.R
\name{bootstrap_k}
\alias{bootstrap_k}
\title{Compute bootstrap (internal)}
\usage{
bootstrap_k(rgcca_res, ...)
}
\arguments{
\item{rgcca_res}{Result of rgcca function}

\item{...}{Further graphical parameters (see plot2D)}
}
\value{
A list of RGCCA bootstrap weights
}
\description{
Internal function for computing boostrap of RGCCA
}
