% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_k.R
\name{bootstrap_k}
\alias{bootstrap_k}
\title{Compute bootstrap (internal)}
\usage{
bootstrap_k(rgcca, blocks = NULL, connection = 1 -
  diag(length(blocks)), tau = rep(1, length(blocks)), ncomp = rep(2,
  length(blocks)), scheme = "factorial", init = "svd", bias = TRUE,
  tol = 0.001, type = "rgcca", superblock = TRUE)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{blocks}{A list of matrix}

\item{connection}{A matrix giving the connection between the blocks}

\item{tau}{A vector of float (or character for 'optimal' setting) giving the
shrinkage parameter for covariance maximization}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{scheme}{A character giving the link function for covariance maximization}

\item{init}{A character among "svd" (Singular Value Decompostion) or "random"
for algorithm initialization}

\item{bias}{A boolean for a biased variance estimator}

\item{tol}{An integer for the stopping value for convergence}

\item{type}{A character giving the type of analysis: c('rgcca', 'cpca-w', 'gcca', 'hpca', 'maxbet-b', 'maxbet', 'maxdiff-b','maxdiff', 'maxvar-a', 'maxvar-b', 'maxvar', 'niles', 'r-maxvar', 'rcon-pca',
'ridge-gca', 'sabscor', 'ssqcor', 'ssqcor', 'ssqcov-1', 'ssqcov-2', 'ssqcov',
'sum-pca', 'sumcor', 'sumcov-1', 'sumcov-2', 'sumcov.', 'sabscov', 'plspm','cca', 'ra', 'ifa', 'pls','pca')}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}
}
\value{
A list of RGCCA bootstrap weights
}
\description{
Internal function for computing boostrap of RGCCA
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks)
bootstrap_k(rgcca_out)
bootstrap_k(rgcca_out, lapply(blocks, scale), superblock = FALSE)
}
