% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quantitative.R
\name{check_quantitative}
\alias{check_quantitative}
\title{Check if a dataframe contains no qualitative variables}
\usage{
check_quantitative(df, fo, h = FALSE)
}
\arguments{
\item{df}{A dataframe or a matrix}

\item{fo}{A character giving the name of the tested file}

\item{h}{A bolean giving either the presence (TRUE) or absence (FALSE) of
a header}
}
\description{
Check if a dataframe contains no qualitative variables
}
\examples{
df = matrix(runif(20), 10, 2)
check_quantitative(df, 'data')
\dontrun{
df[,2] = LETTERS[seq(10)]
check_quantitative(df, 'data', TRUE)
# Error
}
}
