% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{commonRow}
\alias{commonRow}
\title{Get the common row names}
\usage{
commonRow(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A vector of character with the common rownames
}
\description{
Get the common row names
}
