% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.r
\name{comparison}
\alias{comparison}
\title{comparison of two RGCCA results}
\usage{
comparison(rgcca1, rgcca2, nAxe = 1, selec = 10,
  selectPatient = NULL, indNA = NULL)
}
\arguments{
\item{rgcca1}{A result of a RGCCA function}

\item{rgcca2}{Another result of RGCCA function}

\item{selec}{Number of biomarkers to be selected}

\item{selectPatient}{A vector allowing to select only some patients for the RV calculation}

\item{nAxe=1}{Number of axes taken into account during the comparison}
}
\value{
\item{A}{A list containing: a: the correlation between axes, rv: the rv coefficient, bm the biomarkers}

\item{crit}{Convergence criterion : abs(1-obj_k/obj_{k-1})}

\item{obj}{Vector containing the mean square error between the predict values and the original non missing values at each iteration}
}
\description{
Compares 2 RGCCA
}
\examples{
 data();...
}
