% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Check that the network is fully connected}
\usage{
correct(x)
}
\arguments{
\item{x}{Design matrix, a symmetric matrix with}
}
\value{
A logical value if it is fully connected or not.
}
\description{
Given the design matrix, checks that all the blocks are connected between them
}
\examples{
connections_wrong <- matrix(c(0, 1, 0, 1, 0, 0, 0, 0, 0), ncol = 3, nrow = 3)
connections_well <- matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), ncol = 3, nrow = 3)
connections_wrong # No connection of block 3 with any other
correct(connections_wrong)
connections_well
correct(connections_well)
connections_wrong[3, 3] <- 1 # Connecting block 3 with itself
correct(connections_wrong) # Still wrong!
}
\references{
\url{https://math.stackexchange.com/a/551947}
}
\author{
Lluís Revilla Sancho
}
