% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNA.r
\name{createNA}
\alias{createNA}
\title{Create a list with missing data (for simulation)}
\usage{
createNA(A, option = "block", pNA = 0.1, nAllRespondants = 4,
  output = "list", seed = NULL)
}
\arguments{
\item{A}{A list of complete matrices with the same number of lines}

\item{option}{"bloc" if the structure of missing data is by block or "ponc" if it is random across variables}

\item{pNA}{A logical value. If scale = TRUE, each column is transformed to have unit variance (default = TRUE).}

\item{nAllRespondants}{If the option "ponc" is chosen,nAllRespondant corresponds to the number of complete individuals required}

\item{output}{list by default, else, can return the concatened list.}

\item{seed}{if NULL, the creation of missing data is totally random (and consequently unreproducible), if numeric value, select randomness according to a seed (and reproducible).}
}
\value{
\item{A}{a list of NA}
}
\description{
Create a dataset with missing values
}
