% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_resampling.R
\name{generate_resampling}
\alias{generate_resampling}
\title{Generate bootstrap samples.}
\usage{
generate_resampling(
  rgcca_res,
  n_boot,
  balanced = TRUE,
  keep_all_variables = FALSE,
  pval = 1e-15,
  verbose = TRUE,
  stratified = F,
  groups = NULL
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{n_boot}{Number of bootstrap samples. Default: 100.}

\item{balanced}{A boolean indicating if a balanced bootstrap procedure is
performed or not (default is TRUE).}

\item{keep_all_variables}{A boolean indicating if all variables have to be
kept even when some of them have null variance for at least one bootstrap
sample (default is FALSE).}

\item{pval}{For all the variables, a threshold for the proportion of the most
frequent value of this variable is computed. This threshold is evaluated so
that the probability to sample only this value is below `pval`. This
probability is corrected by the number of bootstrap samples and the number
of variables (default is `1e-15`).}

\item{verbose}{Logical value indicating if the progress of the
bootstrap procedure is reported.}

\item{stratified}{Logical. Should stratified sampling be performed, i.e. 
the proportion of observations from each class in the learning sets be the 
same as in the whole data set ?}

\item{groups}{A vector of class labels, either numeric or a factor. 
Must be given if stratified = TRUE.}
}
\value{
\item{full_idx}{A list of size n_boot containing the observations
kept for each bootstrap sample.}

\item{sd_null}{A list of size the number of block
containing the variables that were removed from each block for all the
bootstrap samples. Variables are removed if they appear to be of null variance
in at least one bootstrap sample. If no variable is removed, return NULL.}
}
\description{
Generate `n_boot` bootstrap samples.
}
\details{
The bootstrap samples are generated so that it is very unlikely that they
will be associated with a 0 variance variable. However such situation can
appear in very rare cases.

The simplest procedure to avoid such situation would be to compute the
variance of each variable in each bootstrap sample. However, in
high-dimensional cases, this is time consuming. Thus, the function starts
by identifying variables that present a risk of being of null variance in
at least one bootstrap sample. A variable is detected as such if the
probability of sampling `N` times (where `N` is the number of observations)
its most frequent observed value is higher than a specific threshold
named `pval` (by default set to `1e-15`) corrected by the number of bootstrap
sample `n_boot` and the number of variables in the whole data-set
(\eqn{\sum_{j=1}^Jp_{j}}). In the end, a variable is defined as risky if the
proportion of its most frequent observed value is strictly higher than:
\deqn{risky\textunderscore threshold = \left(\frac{pval}{n_{boot}*
\left(\sum_{j=1}^J p_{j}\right)}\right)^{1/N}.}
This value is necessarily lower than `1` as `pval` is lower than `1`.
However, it could be lower than \eqn{1/N}, which means that all variables
are defined as risky. That is why the maximum value between
\eqn{risky\textunderscore threshold} and \eqn{1/N} is taken.

Once risky variables have been identified, the function computes the
variance of each risky variable in each bootstrap sample. If there isn't any
0 variance risky variable, the bootstrap samples are returned as is.
Otherwise, three cases are possible:
\itemize{
\item `keep_all_variables = F`, then the risky variables that are of 0
variance in at least one bootstrap sample are removed. If they correspond to
a whole block, an error message is generated.
\item `keep_all_variables = T`, two possibilities :
\itemize{
\item If `balance = T`, the procedure is repeated at most `5` times until all
bootstrap samples do not present any 0 variance variable.
\item If `balance = F`, an heuristic procedure is used to modify the sampling
probability of each observation in order to keep all variables.
}
}

A short detail of this lastly mentioned heuristic consist in replacing each
observed value of the risky variables by \eqn{1-\frac{N_k}{N}}
(where \eqn{N_k} corresponds to the number of times this value is observed),
normalized so that the sum through all the observations (for each variable)
equals `1`. Then, if \eqn{prob_i} defines the sampling probability for the
observation associated with line \eqn{i}, it is defined as the maximum value
of the previous matrix through all risky variables (again normalized so that
\eqn{\sum_{i=1}^N prob_i = 1}). Thus the sampling probability of an
observation is more of less associated with `1 - its proportion in the
variable where this observation is in the lowest frequent group
(through all risky variables)`.
}
\keyword{internal}
