% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap.R
\name{get_bootstrap}
\alias{get_bootstrap}
\title{Extract statistics from a fitted bootstrap object}
\usage{
get_bootstrap(
  b,
  type = "weight",
  comp = 1,
  block = length(b$bootstrap[[1]][[1]]),
  display_order = TRUE,
  adj.method = "fdr"
)
}
\arguments{
\item{b}{A fitted bootstrap object (see  \code{\link[RGCCA]{bootstrap}})}

\item{type}{Character string indicating the bootstrapped object to print: 
block-weight vectors ("weight", default) or block-loading vectors 
("loadings").}

\item{comp}{An integer indicating the block-weight/loading vector to plot.}

\item{block}{A numeric vector indicating the blocks to consider.}

\item{display_order}{A logical value to display the order of the variables}

\item{adj.method}{character string indicating the method used for p-value
adjustment (default: fdr - a.k.a Benjamini-Hochberg correction)}
}
\value{
A dataframe containing:
\itemize{
\item 'mean' for the mean of the bootstrap weights/loadings (non-null for SGCCA)
\item 'estimate' for RGCCA block-weight/block-loading vectors
\item 'sd' for the bootstrap estimate of the standard error of the 
(non-null in case of SGCCA) bootstrap weights/loadings
\item 'lower/upper_bound' for the lower and upper intervals 
(0.025/0.975 percentile).
\item 'bootstrap_ratio' defined as the ratio between weight/loadings estimate
and the bootstrap estimate of the standard deviation.
\item 'pval' for p-value (see details)
\item 'adjust.pval' for adjusted p-value (default value: fdr (Benjamini-
Hochberg correction))
\item 'occurrences' for non-zero occurences (for SGCCA)
\item 'sign' for significant ('*') or not ('ns') p-value (alpha = 0.05)
(see details)
}
}
\description{
This function extracts statistical information from a fitted bootstrap
object (see \code{\link[RGCCA]{bootstrap}}).
}
\details{
For RGCCA, the p-value is computed by assuming that the ratio of the blocks
weight values to the bootstrap estimate of the standard deviation follows
the standardized normal distribution.
By including sparsity (with "sgcca","spls" or "spca"), the frequency of a
selected variable may depend on both the level of sparsity and the total
number of variables in each block.
For a random selection of the variable within the block, the number of
occurrences (0 or 1) follows a Bernoulli distribution with the parameter
p = proportion of selected variables in the block.
This proportion is estimated by the average number of selected variables
over all bootstraps divided by the total number of variables in each block
(p_j). On a larger number of bootstrap samples, the number of occurrences
follows a binomial distribution B(n,p) with n=number of bootstraps.
The test is based on the following null hypothesis: "the variable is
randomly selected according to B(n,p)". This hypothesis is rejected when
the number of occurrences is higher than the 1-(0.05/p_j)th quantile
}
\examples{
# Bootstrap confidence intervals and p-values for RGCCA
data("Russett")
blocks = list(agriculture = Russett[, seq(3)],
              industry = Russett[, 4:5],
              politic = Russett[, 6:11])

rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, n_boot = 5, n_cores = 1)
get_bootstrap(boot, type = "loadings")

# Stability of the selected variables for SGCCA

}
\seealso{
\code{\link[RGCCA]{bootstrap}},
\code{\link[RGCCA]{plot.bootstrap}},
\code{\link[RGCCA]{print.bootstrap}}
}
