% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_all.R
\name{get_cor_all}
\alias{get_cor_all}
\title{Correlation between the blocks}
\usage{
get_cor_all(rgcca_res, blocks = rgcca_res$call$blocks,
  comps = get_comp_all(rgcca_res))
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{blocks}{A list of matrix}

\item{comps}{A matrix containg the components of all the blocks}
}
\value{
A list of matrix containg the pairwise correlation of the blocks
}
\description{
Calculates the pairwise correlation between each block for each of the 
components of the analysis
}
\examples{
coord = lapply(seq(3),
   function(x) matrix(runif(15 * 2, min = -1), 15, 2))
AVE_X = lapply(seq(3), function(x) runif(2))
for (i in 1:length(coord))
row.names(coord[[i]]) = seq(15)
rgcca_out = list(Y = coord, AVE = list(AVE_X = AVE_X), 
call = list(blocks = coord, ncomp = rep(2,3))) # TODO
# Using a superblock
resp = as.matrix(rep(LETTERS[seq(3)], each = 5))
row.names(resp) = seq(15)
rgcca_out$call$type="rgcca"
class(rgcca_out) = "rgcca"
plot_ind(rgcca_out, resp)
# Using the first block
resp = as.matrix(runif(15, min=-15, max = 15))
row.names(resp) = seq(15)
plot_ind(rgcca_out, resp, 1, 2, 1)
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")])
X_ind = as.matrix(Russett[,c("gnpr","labo")])
X_polit = as.matrix(Russett[ , c("demostab", "dictator")])
A = list(X_agric, X_ind, X_polit)
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
result.rgcca = rgcca(A, C, tau = c(1, 1, 1), scheme = "factorial",
scale = TRUE,ncomp=rep(2,3))
plot_ind(result.rgcca,i_block=1)
}
