% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_all.R
\name{get_cor_all}
\alias{get_cor_all}
\title{Correlation between the blocks}
\usage{
get_cor_all(rgcca, blocks = rgcca$call$blocks,
  comps = get_comp_all(rgcca))
}
\arguments{
\item{blocks}{A list of matrix}

\item{comps}{A matrix containg the components of all the blocks}
}
\value{
A list of matrix containg the pairwise correlation of the blocks
}
\description{
Calculates the pairwise correlation between each block for each of the 
components of the analysis
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
get_cor_all(rgcca_out)
}
