% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctr.R
\name{get_ctr}
\alias{get_ctr}
\title{Variable contribution}
\usage{
get_ctr(
  rgcca_res,
  compx = 1,
  compy = 2,
  compz = NULL,
  i_block = length(rgcca_res$call$blocks),
  type = "cor",
  collapse = FALSE,
  i_block_2 = NULL
)
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{compz}{An integer giving the index of the analysis component used
for the z-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A character giving the choice ot the index between cor or weight}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{i_block_2}{An integer giving the index of a list of blocks to be 
correlated to i_block if this option is selected (default to i_block)}
}
\value{
A dataframe containing the indexes for each selected components
}
\description{
Extract the contibution of variables to the model by using correlation or weight
}
