% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctr2.R
\name{get_ctr2}
\alias{get_ctr2}
\title{Get the indexes of the analysis}
\usage{
get_ctr2(rgcca_res, compx = 1, compy = 2, compz = NULL,
  i_block = length(rgcca_res$call$blocks), type = "cor",
  n_mark = 100, collapse = FALSE, remove_var = TRUE, resp = NULL)
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{compz}{An integer giving the index of the analysis component used
for the z-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A character giving the choice ot the index between cor or weight}

\item{n_mark}{An integer giving the number of top variables to select}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}
}
\value{
A matrix containg the indexes (correlation of the blocks with a 
component or their weights) for each selected component and an associated response
}
\description{
Get the indexes of the analysis
}
