% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges.R
\name{get_edges}
\alias{get_edges}
\title{Creates the edges for a design matrix}
\usage{
get_edges(rgcca)
}
\value{
A dataframe with tuples of connected rgcca$call$blocks
}
\description{
Creates the edges for a design matrix
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
get_edges(rgcca_out)
}
