% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Creates the nodes for a design matrix}
\usage{
get_nodes(rgcca_res)
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}
}
\value{
A dataframe with rgcca_res$call$blocks in rows and the number of variables, of rows
and tau or sparsity in columns
}
\description{
Creates the nodes for a design matrix
}
