% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Creates the nodes for a design matrix}
\usage{
get_nodes(rgcca, tau = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{tau}{A vector of float (or character for 'optimal' setting) giving the
shrinkage parameter for covariance maximization}
}
\value{
A dataframe with rgcca$call$blocks in rows and the number of variables, of rows
and tau or c1 in columns
}
\description{
Creates the nodes for a design matrix
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
get_nodes(rgcca = rgcca_out)
}
