% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getBlocsVariables}
\alias{getBlocsVariables}
\title{Get Blocs Variables}
\usage{
getBlocsVariables(rgcca)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}
}
\description{
Get a vector of block name for each corresponding variable. The last block is considered as the superblock and ignored.
}
\examples{
rgcca.res = list(a = rep(NA, 4))
names(rgcca.res$a) = letters[1:4]
getBlocsVariables(rgcca.res)
# a, b, c
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
