% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeColmeans.R
\name{imputeColmeans}
\alias{imputeColmeans}
\title{Product for Matrices with missing data (pm)}
\usage{
imputeColmeans(A)
}
\arguments{
\item{A}{A list of matrices}
}
\value{
List of the matrices imputed by colmeans
}
\description{
Impute a list of matrices by the colmeans
}
