% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{keepCommonRow}
\alias{keepCommonRow}
\title{Keep only the rows with the same names among a list of dataframe}
\usage{
keepCommonRow(list_m)
}
\arguments{
\item{list_m}{A list of dataframe}
}
\value{
A list of dataframe
}
\description{
Keep only the rows with the same names among a list of dataframe
}
