% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_blocks.R
\name{load_blocks}
\alias{load_blocks}
\title{Create a list of matrix from loading files corresponding to blocks}
\usage{
load_blocks(
  file,
  names = NULL,
  separator = "\\t",
  header = TRUE,
  rownames = 1,
  decimal = "."
)
}
\arguments{
\item{file}{A character giving the path of a file}

\item{names}{A character giving a list of names for the blocks}

\item{separator}{A character giving the column separator}

\item{header}{A bolean giving the presence or the absence of the header}

\item{rownames}{An integer corresponding to the column number of the rownames (NULL otherwise)}

\item{decimal}{A character giving the decimal separator}
}
\value{
A list matrix corresponding to the blocks
}
\description{
Create a list of matrix from loading files corresponding to blocks
}
\examples{
\dontrun{
load_blocks (TRUE,
    "inst/extdata/agriculture.tsv,inst/extdata/industry.tsv,inst/extdata/politic.tsv",
    "agric,ind,polit")
}
}
