% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_text.R
\name{load_file_text}
\alias{load_file_text}
\title{Creates a matrix from loading a file}
\usage{
load_file_text(file, sep = "\\t", rownames = 1, header = TRUE,
  one_column = FALSE, dec = ".")
}
\arguments{
\item{file}{A character giving the path of a file}

\item{sep}{A character giving the column separator}

\item{rownames}{An integer corresponding to the column number of the row
names (NULL otherwise)}

\item{header}{A bolean giving the presence or the absence of the header}

\item{one_column}{A boolean for a file with  an only-one column}
}
\value{
A matrix containing the loaded file
}
\description{
Creates a matrix from loading a file
}
\examples{
\dontrun{
load_file_text('data/agriculture.tsv')
}
}
