% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{loadExcel}
\alias{loadExcel}
\title{Creates a data frame from loading a file}
\usage{
loadExcel(f, sheet, row.names = 1, h = TRUE)
}
\arguments{
\item{f}{A character giving the file name}

\item{sheet}{A character giving the sheet name}

\item{row.names}{A vector of characters giving the names of the rows}

\item{h}{A bolean giving the presence or the absence of the header}
}
\value{
A matrix containing the loaded file
}
\description{
Creates a data frame from loading a file
}
\examples{
\dontrun{
loadExcel("data/blocks.xlsx", "industry")
}
}
