% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIRGCCA.R
\name{MIRGCCA}
\alias{MIRGCCA}
\title{MIRGCCA: Multiple imputation for RGCCA}
\usage{
MIRGCCA(blocks, option = "knn", superblock = TRUE, k = 5, ni = 5,
  scale = TRUE, sameBlockWeight = TRUE, tau = rep(1:length(A)),
  klim = NULL, output = "mean", scheme = "centroid", tol = 1e-08,
  connection = NULL, ncomp = rep(2, length(A)), naxis = 1)
}
\arguments{
\item{blocks}{A list of matrix}

\item{option}{"knn" for k Nearest Neigbors or "em" for Expectation Maximization}

\item{superblock}{TRUE if the A list should be returned in the output, FALSE ifelse}

\item{k}{Integer representing the number of neighbors or "auto" or "all"}

\item{ni}{number of imputations}

\item{scale}{A boolean scaling the blocks}

\item{sameBlockWeight}{If TRUE, all blocks are weighted by their own variance: all the blocks have the same weight}

\item{tau}{tau is either a \eqn{1 \times J} vector or a \eqn{\mathrm{max}(ncomp) \times J} matrix, and contains the values 
of the shrinkage parameters (default: tau = 1, for each block and each dimension).
If tau = "optimal" the shrinkage paramaters are estimated for each block and each dimension using the Schafer and Strimmer (2005)
analytical formula . If tau is a \eqn{1\times J} numeric vector, tau[j] is identical across the dimensions of block \eqn{\mathbf{X}_j}. 
If tau is a matrix, tau[k, j] is associated with \eqn{\mathbf{X}_{jk}} (\eqn{k}th residual matrix for block \eqn{j})}

\item{klim}{TRUE if the A list should be returned in the output, FALSE ifelse}

\item{output}{TRUE if the A list should be returned in the output, FALSE ifelse}

\item{scheme}{A character giving the link function for covariance maximization}

\item{tol}{An integer for the stopping value for convergence}

\item{connection}{A matrix giving the connection between the blocks}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{naxis}{number of component to select in the superblock for the estimation of missing data for "em" option}
}
\value{
\item{rgcca0}{RGCCA results for the reference dataset}

\item{data}{list of imputed data obtained}

\item{rgccaList}{list of RGCCA obtained}
}
\description{
Multiple imputation for RGCCA
}
\details{
This method allows multiple imputation for RGCCA with several options.
}
\examples{
set.seed(42);X1=matrix(rnorm(500),100,5);
set.seed(22);X2=matrix(rnorm(400),100,4);
set.seed(2);X3=matrix(rnorm(700),100,7);
rownames(X1)=rownames(X2)=rownames(X3)=paste("S",1:100,sep="")
colnames(X1)=paste("A",1:5)
colnames(X2)=paste("B",1:4)
colnames(X3)=paste("C",1:7)
X1[1,]=NA
X2[7,1]=NA
 X2[5,1]=NA
 X3[3,1:2]=NA
 A=list(X1,X2,X3)
res=MIRGCCA(A,k=3,ni=5,scale=TRUE,sameBlockWeight=TRUE,tau=rep(0,3))
}
