% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelize.R
\name{parallelize}
\alias{parallelize}
<<<<<<< HEAD
\title{Set a list of sockets for parralel package}
=======
\title{parallelize
Set a list of ocket for parralel package in Windows}
>>>>>>> develop
\usage{
parallelize(varlist = c(), nperm, f, n_cores = NULL,
  envir = environment(), applyFunc = "parSapply")
}
\arguments{
\item{varlist}{character vector of names of objects to export}

\item{nperm}{a vector object for a lapply type function}

\item{f}{a function to parralelize}

<<<<<<< HEAD
\item{n_cores}{An integer for the number of cores used in parallelization}

\item{envir}{environment}

\item{applyFunc}{function to be applied}
}
\description{
Set a list of sockets for parralel package
=======
\item{n_cores}{: number of cores to be used}

\item{envir}{: environment}

\item{applyFunc}{: function to be applied}
}
\description{
parallelize
Set a list of ocket for parralel package in Windows
>>>>>>> develop
}
