% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bootstrap.r
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot a fitted bootstrap object}
\usage{
\method{plot}{bootstrap}(
  x,
  block = length(x$rgcca$call$blocks),
  comp = 1,
  type = "weight",
  empirical = TRUE,
  n_mark = 30,
  display_order = TRUE,
  colors = NULL,
  title = NULL,
  cex = 1,
  cex_main = 14,
  cex_sub = 12,
  cex_point = 10,
  cex_lab = 10,
  cex_axis = 10,
  ...
)
}
\arguments{
\item{x}{A fitted bootstrap object (see \code{\link[RGCCA]{bootstrap}})}

\item{block}{A numeric vector indicating the blocks to consider.}

\item{comp}{An integer indicating the block-weight/loading vector to plot.}

\item{type}{Character string indicating the bootstrapped object to plot:
block-weight vectors ("weight", default) or block-loading vectors
("loadings").}

\item{empirical}{A logical value indicating if the bootstrap confidence
intervals and p-values are derived from the empirical distribution.
(defaut: TRUE)}

\item{n_mark}{The number of top variables to display in the correlation
circle}

\item{display_order}{A logical value for ordering the variables}

\item{colors}{Either a vector of integers (each integer corresponding to a
color) or of characters corresponding to names of colors (as "blue",see
colors()) or RGB code ("#FFFFFF").}

\item{title}{A character giving the title of the plot}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{cex_axis}{An integer for the size of the axis text}

\item{...}{Further graphical parameters (see plot2D)}
}
\description{
Plot the results of a fitted bootstrap object.
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)],
              industry = Russett[, 4:5],
              politic = Russett[, 6:11])
fit.rgcca = rgcca(blocks, ncomp = 2, method = "rgcca", tau = 1)
fit.boot = bootstrap(fit.rgcca, n_boot = 20, n_cores = 1)
plot(fit.boot, type = "weight", block = 1, comp=1)
}
