% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.df_bootstrap.r
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{plot.df_bootstrap}
\usage{
\method{plot}{bootstrap}(x, type = "1D",
  i_block = length(x$rgcca$call$blocks), colors = NULL, title = NULL,
  cex = 1, ...)
}
\arguments{
\item{x}{result of get_bootstrap}

\item{type}{"1D" or "2D"}

\item{i_block}{An integer giving the index of a list of blocks}

\item{colors}{reoresenting a vector of colors}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{cex}{An integer for the size of the plot parameters}

\item{...}{Other parameters (see plot_histogram)}
}
\description{
plot.df_bootstrap
}
