% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.df_bootstrap.r
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot a bootstrap object}
\usage{
\method{plot}{bootstrap}(
  x,
  type = "1D",
  block = length(x$rgcca$call$blocks),
  comp = 1,
  n_mark = 30,
  bars = "quantile",
  colors = NULL,
  title = NULL,
  cex = 1,
  n_cores = parallel::detectCores() - 1,
  collapse = FALSE,
  cex_main = 14,
  cex_sub = 12,
  cex_point = 10,
  cex_lab = 10,
  cex_axis = 10,
  ...
)
}
\arguments{
\item{x}{A bootstrap object (see \code{\link[RGCCA]{bootstrap}})}

\item{type}{A character to select a type of plot among "1D" or "2D"}

\item{block}{A vector of integers giving the index of a list of blocks or a block = 'all' for all of them}

\item{comp}{An integer giving the index of the analysis components}

\item{n_mark}{An integer giving the number of top variables to select}

\item{bars}{A character giving the variability among "sd" (standard deviations bars) , "stderr"(bars of standard deviation divided by sqrt(n)) or "quantile" (for the 0.05-0.95 quantiles bars)}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{title}{A character giving the title of the plot}

\item{cex}{An integer for the size of the plot parameters}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{cex_axis}{An integer for the size of the axis text}

\item{...}{Further graphical parameters (see plot2D)}
}
\description{
Plot the results of a bootstrap object. The representation can be with bars
(1D) or a biplot (2D) (see details).
}
\details{
\itemize{
\item type = '1D': barplot of the best variables from a bootstrap with, on the x-axis,
the number of non-zero occurrences (SGCCA) or the mean of the bootstrap weights 
(RGCCA). The bars are colored according to the significant 95% bootstrap 
intervals ('*' or 'ns'; see 'p.vals' in details of \code{\link[RGCCA]{get_bootstrap}}). In SGCA, the significant variables 
 are those above the three bars, respectively, with an alpha = 0.05 
 (dark red), 0.01 (red) and 0.001 (light red).
\item type = '2D' : graph of the best variables from a bootstrap with, in x-axis, the number of
non-zero occurences (SGCCA) or the significant 95% bootstrap 
intervals (RGCCA). In in y-axis are the bootstrap-ratios (mean/sd) 
Negative weights are colored in red and the positive ones are in green.
}
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
b=bootstrap(rgcca_out, n_boot = 2, n_cores = 1)
plot(b,n_cores=1)
}
