% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bootstrap.r
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot a fitted bootstrap object}
\usage{
\method{plot}{bootstrap}(
  x,
  block = seq_along(x$rgcca$call$blocks),
  comp = 1,
  type = "weights",
  empirical = TRUE,
  n_mark = 30,
  display_order = TRUE,
  show_sign = TRUE,
  title = NULL,
  cex = 1,
  cex_sub = 12 * cex,
  cex_main = 14 * cex,
  cex_lab = 12 * cex,
  cex_point = 3 * cex,
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{A fitted bootstrap object (see \code{\link[RGCCA]{bootstrap}})}

\item{block}{A numeric corresponding to the block(s) to plot.}

\item{comp}{A numeric vector indicating the components to consider.}

\item{type}{Character string indicating the bootstrapped object to plot:
block-weight vectors ("weights", default) or block-loading vectors
("loadings").}

\item{empirical}{A logical value indicating if the bootstrap confidence
intervals and p-values are derived from the empirical distribution.
(defaut: TRUE)}

\item{n_mark}{An integer defining the maximum number of variables plotted.}

\item{display_order}{A logical value for ordering the variables.}

\item{show_sign}{A logical for showing significance levels.}

\item{title}{A character string giving the title of the plot.}

\item{cex}{A numeric defining the size of the objects in the plot. Default
is one.}

\item{cex_sub}{A numeric defining the font size of the subtitle. Default is
12 * cex.}

\item{cex_main}{A numeric defining the font size of the title. Default is
14 * cex.}

\item{cex_lab}{A numeric defining the font size of the labels. Default is
12 * cex.}

\item{cex_point}{A numeric defining the font size of the points. Default is
3 * cex.}

\item{colors}{Colors used in the plots. Default is based on the
colorblind-friendly palette recommended in ggplot.}

\item{...}{additional graphical parameters}
}
\description{
Plot the results of a fitted bootstrap object. Each block variable is shown
along with its associated bootstrap confidence interval and stars reflecting
the p-value of assigning a strictly positive or negative weight to this
block variable.
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
fit.rgcca <- rgcca(blocks, ncomp = 2, method = "rgcca", tau = 1)
fit.boot <- bootstrap(fit.rgcca, n_boot = 20, n_cores = 1)
plot(fit.boot, type = "weight", block = 1, comp = 1)
}
