% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bootstrap.r
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot a fitted bootstrap object}
\usage{
\method{plot}{bootstrap}(
  x,
  block = length(x$rgcca$call$blocks),
  comp = 1,
  n_mark = 30,
  bars = "quantile",
  colors = NULL,
  title = NULL,
  cex = 1,
  cex_main = 14,
  cex_sub = 12,
  cex_point = 10,
  cex_lab = 10,
  cex_axis = 10,
  ...
)
}
\arguments{
\item{x}{A fitted bootstrap object (see \code{\link[RGCCA]{bootstrap}})}

\item{block}{A vector indicating the blocks to consider.}

\item{comp}{An integer giving the index of the analysis components}

\item{n_mark}{The number of top variables to display in the correlation 
circle}

\item{bars}{Character string giving the variability among "sd" (error bar of
standard deviation) , "stderr" (error bar of standard deviation divided by 
sqrt(n)) or "quantile" (error bar of 0.05/0.95-quantiles)}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{title}{A character giving the title of the plot}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{cex_axis}{An integer for the size of the axis text}

\item{...}{Further graphical parameters (see plot2D)}
}
\description{
Plot the results of a fitted bootstrap object.
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], 
              industry = Russett[, 4:5], 
              politic = Russett[, 6:11])
fit.rgcca = rgcca(blocks)
fit.boot=bootstrap(fit.rgcca, n_boot = 2, n_cores = 1)
plot(fit.boot)
}
