% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.R
\name{plot.cv}
\alias{plot.cv}
\title{plot.cv}
\usage{
\method{plot}{cv}(x, ...)
}
\arguments{
\item{x}{A rgcca_crossvalidation object (see \code{\link[RGCCA]{rgcca_crossvalidation}})}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
Plot the results of rgcca_cv object in the space of the samples. 
The actual values of the GCCA (in red), could be contrasted with those 
predicted by the validation (in green).
}
\examples{
data(Russett)
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
             politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, response = 3,ncomp=2)
cv=rgcca_crossvalidation(rgcca_out, validation = "kfold", k = 5, n_cores = 1)
plot(cv)
}
