% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.R
\name{plot.cv}
\alias{plot.cv}
\title{plot.cv}
\usage{
\method{plot}{cv}(x, type = "ind", ...)
}
\arguments{
\item{x}{result of rgcca_crossvalidation (see  \code{\link[RGCCA]{rgcca_crossvalidation}} )}

\item{type}{"ind"}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
Plots the results of rgcca_cv (in abscissa, the configurations, in ordinates the RMSE on test dataset)
}
\examples{
data(Russett)
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
             politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, response = 3)
cv=rgcca_crossvalidation(rgcca_out, validation = "kfold", k = 5, n_cores = 1)
plot(cv)
}
