% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cval.r
\name{plot.cval}
\alias{plot.cval}
\title{plot.cval}
\usage{
\method{plot}{cval}(x, bars = "sd", alpha = 0.05, cex = 1,
  cex_main = 14 * cex, cex_sub = 10 * cex, ...)
}
\arguments{
\item{x}{A rgcca_cv object (see \link{rgcca_cv})}

\item{bars}{A character among "sd" for standard deviations, "stderr" for standard error (standard deviations divided by sqrt(n), "ci" for confidence interal, "cim" for confidence interval of the mean.}

\item{alpha}{A numeric value giving the risk for the confidence interval bars (ci or cim).}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{...}{Further plot options}
}
\description{
plot.cval
}
\examples{
data("Russett")
blocks <- list(
    agriculture = Russett[, seq(3)],
    industry = Russett[, 4:5],
    politic = Russett[, 6:11])
    res=rgcca_cv(blocks, response=3,type="rgcca",par="tau",par_value=c(0,0.2,0.3),n_cv=1,n_cores=1)
   plot(res)
}
