% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cval.R
\name{plot.cval}
\alias{plot.cval}
\title{Plot cross-validation}
\usage{
\method{plot}{cval}(
  x,
  type = "sd",
  cex = 1,
  cex_main = 14 * cex,
  cex_sub = 12 * cex,
  cex_point = 3 * cex,
  cex_lab = 12 * cex,
  display_order = TRUE,
  ...
)
}
\arguments{
\item{x}{A rgcca_cv object (see \link{rgcca_cv})}

\item{type}{Character string indicating the statistics in the box plots:
mean plus or minus standard deviation ("sd", default),
\itemize{
\item "sd" (default): the middle bar corresponds to the mean and limits of
the boxes are given by the mean plus or minus the standard deviation.
\item "stderr": the middle bar corresponds to the mean and limits of
the boxes are given by the mean plus or minus the standard deviation divided
by the square roots of the number of folds.
\item "quantile": the middle bar corresponds to the median and limits of
the boxes are given by the 25% and 75% quantiles.
\item "points": box plots are removed and only the points are kept.}}

\item{cex}{A numeric defining the size of the objects in the plot. Default
is one.}

\item{cex_main}{A numeric defining the font size of the title. Default is
14 * cex.}

\item{cex_sub}{A numeric defining the font size of the subtitle. Default is
12 * cex.}

\item{cex_point}{A numeric defining the font size of the points. Default is
3 * cex.}

\item{cex_lab}{A numeric defining the font size of the labels. Default is
12 * cex.}

\item{display_order}{A logical value for ordering the variables.}

\item{...}{additional graphical parameters}
}
\description{
Plot a cross-validation object (tuning RGCA parameters in 'supervised' mode).
The parameters tuned for maximizing RMSE is displayed in the title. In
the x-axis, the tuning parameter set. In the y-axis, the average of the
cross-validation score. The best parameters are in red by default.
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:8]
)
res <- rgcca_cv(blocks,
  response = 3, method = "rgcca", par_type = "tau",
  par_value = c(0, 0.2, 0.3), n_run = 1, n_cores = 1
)
plot(res)
}
