% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.list_rgcca.R
\name{plot.list_rgcca}
\alias{plot.list_rgcca}
\title{plots a list_rgcca object}
\usage{
\method{plot}{list_rgcca}(x, type = "ind", resp = rep(1,
  NROW(x$Y[[1]])), i_block = 1, i_block_y = i_block, compx = 1,
  compy = 2, remove_var = FALSE, text_var = TRUE, text_ind = TRUE,
  response_name = "Response", no_overlap = FALSE, title = NULL,
  n_mark = 100, collapse = FALSE, cex = 1, cex_sub = 10,
  cex_main = 14, cex_lab = 12, colors = NULL, errorbar = "ci", ...)
}
\arguments{
\item{x}{Result of MIRGCCA function (see\code{\link[RGCCA]{MIRGCCA}} ) or a list containing 
\itemize{\item{rgcca0}{ RGCCA results for the reference dataset}
\item{data}{ list of imputed data obtained}
 \item{rgccaList}{ list of RGCCA obtained}}}

\item{type}{Type among c("ind","var","both","ave","cor","weight","network").  See details.}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{i_block}{An integer giving the index of a list of blocks}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{text_var}{A bolean to represent the variables with their row names (TRUE)
or with circles (FALSE)}

\item{text_ind}{A bolean to represent the individuals with their row names (TRUE)
or with circles (FALSE)}

\item{response_name}{A character giving the legend title}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{title}{Character for the title of the plot}

\item{n_mark}{An integer giving the number of top variables to select}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_main}{An integer for the size of the title}

\item{cex_lab}{An integer for the size of the axis titles}

\item{colors}{representing a vector of the colors used in the graph. Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{errorbar}{("CImean","CIscores","sd")}

\item{...}{Further graphical parameters applied to both (individual and variable) spaces}
}
\description{
plots a list_rgcca object
}
\examples{
set.seed(42);X1=matrix(rnorm(500),100,5);
set.seed(22);X2=matrix(rnorm(400),100,4);
set.seed(2);X3=matrix(rnorm(700),100,7);
rownames(X1)=rownames(X2)=rownames(X3)=paste("S",1:100,sep="")
colnames(X1)=paste("A",1:5)
colnames(X2)=paste("B",1:4)
colnames(X3)=paste("C",1:7)
X1[1,]=NA
X2[7,1]=NA
X2[5,1]=NA
X3[3,1:2]=NA
A=list(X1,X2,X3)
res=MIRGCCA(A,k=3,ni=5,scale=TRUE,scale_block=TRUE,tau=rep(1,3))
plot(res,type="ind")
}
