% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.patternNA.r
\name{plot.patternNA}
\alias{plot.patternNA}
\title{plot.patternNA}
\usage{
\method{plot}{patternNA}(x, type = "all", color = "springgreen4",
  legend = FALSE, scale = TRUE, outlierVisible = FALSE, ...)
}
\arguments{
\item{x}{A get_patternNA object (see \code{\link[RGCCA]{get_patternNA}})}

\item{type}{"all" if all blocks should be plotted, ifelse an integer corresponding to the position of the block to be plotted in the initial list}

\item{color}{color of the non-missing data. If rainbow, depends of the values of the non-missing data}

\item{legend}{if TRUE the legend is plotted}

\item{scale}{if TRUE all the variables are scaled before graphical representation}

\item{outlierVisible}{if FALSE, the outliers will be -2* standard deviations if negative, 2 standard deviations if positive}

\item{...}{Further graphical parameters (see \code{\link[graphics]{plot}})}
}
\description{
Plot an object of class patternNA
}
\examples{
X1=matrix(rnorm(150),30,5)
X2=matrix(rnorm(150),30,5)
X3=matrix(rnorm(150),30,5)
X1[1:10,1]=NA
X1[2,2]=NA
X1[11,]=NA
X2[1:10,1]=NA
colnames(X1)=paste0("A",1:5)
colnames(X2)=paste0("B",1:5)
colnames(X3)=paste0("C",1:5)
A=list(bloc1=X1,bloc2=X2,bloc3=X3)
p=get_patternNA(A)
plot(p)
}
\seealso{
\link{get_patternNA},\link{whichNAmethod}
}
