% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca_permutation.R
\name{plot.permutation}
\alias{plot.permutation}
\title{plot.permutation
Plots a permutation object. The parameters tuned for maximizing RGCCA criteria is displayed in the title. 
In x, the index of combination (number corresponding to the rownames of tuning parameters object). In ordinate, a score depending of the type parameter (RGCCA criterion for crit, and zstat for the pseudo z-scores)}
\usage{
\method{plot}{permutation}(x, type = "crit", cex = 1, title = NULL,
  cex_main = 14 * cex, cex_sub = 12 * cex, bars = "points",
  cex_point = 3 * cex, cex_lab = 19 * cex, ...)
}
\arguments{
\item{x}{result of rgcca_permutation (see  \code{\link[RGCCA]{rgcca_permutation}} )}

\item{type}{A character giving the type of the index to look at (among 'crit' for
the RGCCA criterion and 'zstat' for the pseudo Z-score)}

\item{cex}{An integer for the size of the plot parameters}

\item{title}{A character for the title of the plot}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{bars}{A character giving representation of the variability among "points", "stderr" or "sd"}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{...}{Further graphical parameters}
}
\description{
plot.permutation
Plots a permutation object. The parameters tuned for maximizing RGCCA criteria is displayed in the title. 
In x, the index of combination (number corresponding to the rownames of tuning parameters object). In ordinate, a score depending of the type parameter (RGCCA criterion for crit, and zstat for the pseudo z-scores)
}
\examples{
data("Russett")
A = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
perm <- rgcca_permutation(A, nperm = 2, n_cores = 1)
plot(perm)
perm <- rgcca_permutation(A, perm.par = "sparsity", nperm = 5, n_cores = 1)
plot(perm,type="crit")
plot(perm)
}
