% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.permutation.R
\name{plot.permutation}
\alias{plot.permutation}
\title{Plot fitted rgcca permutation object}
\usage{
\method{plot}{permutation}(
  x,
  type = "crit",
  cex = 1,
  title = NULL,
  cex_main = 14 * cex,
  cex_sub = 12 * cex,
  cex_point = 3 * cex,
  cex_lab = 19 * cex,
  colors = c("red", "grey"),
  display_all = FALSE,
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{A fitted rgcca_permutation object (see
\code{\link[RGCCA]{rgcca_permutation}})}

\item{type}{A character string indicating which criterion to plot
(default is 'crit' for the RGCCA criterion or 'zstat' for the pseudo Z-score)}

\item{cex}{An integer for the size of the plot parameters}

\item{title}{A character giving the title of the plot}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in
the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{colors}{Either a vector of integers (each integer corresponding to a
color) or of characters corresponding to names of colors (as "blue",see
colors()) or RGB code ("#FFFFFF").}

\item{display_all}{A boolean indicating is all combinations have to
be displayed (default is FALSE).}

\item{show_legend}{A boolean indicating if the legend is displayed
(default is TRUE).}

\item{...}{additional graphical parameters}
}
\description{
Plot a fitted rgcca permutation object. The various set of tuning parameters
are represented in the x-axis and the the RGCCA objective function - obtained
from the both the orginal and permuted blocks - in the y-axis. If type =
"zstat" the value of the zstat for the various combinations are reported in
the y-axis.
}
\examples{
data(Russett)
A <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)

perm.out <- rgcca_permutation(A, par_type = "tau", n_perms = 2, n_cores = 1)
print(perm.out)
plot(perm.out)

perm.out <- rgcca_permutation(A,
  par_type = "sparsity",
  n_perms = 5, n_cores = 1
)
print(perm.out)
plot(perm.out, type = "zstat")
}
