% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict.R
\name{plot.predict}
\alias{plot.predict}
\title{plot.predict}
\usage{
\method{plot}{predict}(x, ...)
}
\arguments{
\item{x}{A rgcca_predict object (see \code{\link[RGCCA]{rgcca_predict}})}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
plot.predict
}
\examples{
data(Russett)
}
