% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca.R
\name{plot.rgcca}
\alias{plot.rgcca}
\title{Plot for RGCCA}
\usage{
\method{plot}{rgcca}(
  x,
  type = "weight",
  block = length(x$A),
  comp = 1:2,
  resp = rep(1, NROW(x$Y[[1]])),
  remove_var = FALSE,
  text_var = TRUE,
  text_ind = TRUE,
  response_name = "Response",
  no_overlap = FALSE,
  title = NULL,
  n_mark = 30,
  collapse = FALSE,
  cex = 1,
  cex_sub = 12,
  cex_main = 14,
  cex_lab = 12,
  cex_axis = 10,
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{A RGCCA object (see \code{\link[RGCCA]{rgcca}})}

\item{type}{A character among 'ind', 'var', 'both', 'ave', 'cor', 'weight', 'network' (see details).}

\item{block}{A vector of integers giving the index of a list of blocks or a block = 'all' for all of them}

\item{comp}{An integer giving the index of the analysis components}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{text_var}{A bolean to represent the variables with their row names (TRUE)
or with circles (FALSE)}

\item{text_ind}{A bolean to represent the individuals with their row names (TRUE)
or with circles (FALSE)}

\item{response_name}{A character giving the legend title}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{title}{A character giving the title of the plot}

\item{n_mark}{An integer giving the number of top variables to select}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_main}{An integer for the size of the title}

\item{cex_lab}{An integer for the size of the axis titles}

\item{cex_axis}{An integer for the size of the axis text}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{...}{Further graphical parameters}
}
\description{
Plot different outputs of the results obtained by a rgcca function
}
\details{
\itemize{
 \item "ind" for individual graph: Y of rgcca are plotted. In abscissa Y[[i_block]][,compx], in ordinate, Y[[i_block_y]][,compy]. Each point correspond to one individual. It can be colored with the resp. options. The colors by default can be modified in colors options.
\item  "var" for variable graph: in abscissa, the correlations with the first axis, in ordinate, the correlation with the second axis. 
\item "both": displays both ind and var graph (this requires only one block (i_block=i_block_y) and at least two components in the rgcca calculation (ncomp>1 for this block)
\item "ave": displays the variance average in each block
\item "net": displays the graphical network corresponding to the connection matrix used in the rgcca
\item "cor": barplot corresponding to the correlation of each variable with a chosen axis (specified by i_block and compx). Variables are sorted from the highest to the lowest and only the highest are displayed (to modify the number, use the parameter n_marks)
\item "weight":barplot corresponding to the correlation of each variable with a chosen axis (specified by i_block and compx). Variables are sorted from the highest to the lowest and only the highest are displayed (to modify the number, use the parameter n_marks)
}
}
\examples{
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric, X_ind, X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
resRgcca=rgcca(blocks=A,connection=C,
tau=rep(1,3),ncomp=rep(2,3),superblock=FALSE)
plot(resRgcca,type="cor")
plot(resRgcca,type="weight")
plot(resRgcca,type="ind")
plot(resRgcca,type="var")
plot(resRgcca,type="both")
plot(resRgcca,type="ave")
plot(resRgcca,type="network")
}
