% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca.R
\name{plot.rgcca}
\alias{plot.rgcca}
\title{Plot for RGCCA}
\usage{
\method{plot}{rgcca}(
  x,
  type = "weight",
  block = length(x$blocks),
  comp = 1:2,
  factor = 1,
  rank = 1,
  resp = rep(1, NROW(x$Y[[1]])),
  remove_var = FALSE,
  text_var = TRUE,
  text_ind = TRUE,
  response_name = "Response",
  overlap = TRUE,
  title = NULL,
  n_mark = 30,
  collapse = FALSE,
  cex = 1,
  cex_sub = 12,
  cex_main = 14,
  cex_lab = 12,
  cex_axis = 10,
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{A RGCCA object (see \code{\link[RGCCA]{rgcca}})}

\item{type}{A character among 'ind', 'var', 'both', 'ave', 'cor', 'weight', 
'network', 'factor', 'weight_matrix' (see details).}

\item{block}{A vector indicating the blocks to consider.}

\item{comp}{An integer giving the index of the analysis components}

\item{factor}{An integer giving the index of the mode factor. Only used 
when type is "factor"}

\item{rank}{An integer giving the rank of the factor we want to look at. Only
used when type is "factor"}

\item{resp}{A vector of characters corresponding either to a qualitative
variable with levels or a continuous variable}

\item{remove_var}{logical value. If TRUE, the top 100 variables with the 
highest with the block components are displayed}

\item{text_var}{Boolean value indicating if variable names are plotted 
(default = TRUE)}

\item{text_ind}{boolean value indicating if rownames are plotted 
(default = TRUE).}

\item{response_name}{A character giving the legend title}

\item{overlap}{Boolean value enables avoiding overlapping between labels 
(default = FALSE)}

\item{title}{A character giving the title of the plot}

\item{n_mark}{The number of top variables to display in the correlation 
circle}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_main}{An integer for the size of the title}

\item{cex_lab}{An integer for the size of the axis titles}

\item{cex_axis}{An integer for the size of the axis text}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{...}{additional graphical parameters}
}
\description{
Plot different outputs of the results obtained by a rgcca function
}
\details{
\itemize{
\item "ind" for sample plot. The blocks (block argument) and components 
(comp) that will be used on the horizontal and the vertical axis to plot the 
individuals: (Y[[block[1]]][, comp[1]], Y[[block[2]]][,comp[2]]). Points can 
be colored according to the resp argument. The colors of the points can be 
modified with the colors argument.
\item  "var" for correlation circle.  
first axis, in ordinate, the correlation with the second axis. 
\item "both": displays both sample plot and correlation circle (implemented
only for one block and at least when two components are asked (ncomp >= 2)
\item "ave": displays the average variance explained for each block
\item "net": displays the network of connection between blocks (defined by 
the connection argument) used in the rgcca() function.
\item "cor": barplot of the correlation between variables of 
one block (specified by block) and one of its block components (comp). 
Variables are sorted in decreasing correlations and only the highest 
correlations are displayed. The number of displayed correlations can be set 
with n_marks (defaut value = 30).
\item "weight": barplot of the block weight vector for one 
specific block/component. The weights are sorted from the highest to 
the lowest and only the highest are displayed. The number of displayed 
weights can be set with n_marks
\item "factor": barplot of the block weight factor for one 
specific block/component/mode/rank. The weights are sorted from the highest 
to the lowest and only the highest are displayed. The number of displayed 
weights can be set with n_marks
\item "weight_matrix": heatmap of the weight matrix for 3D tensor blocks.
}
}
\examples{
data(Russett)
status = colnames(Russett)[9:11][apply(Russett[, 9:11], 1, which.max)]
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric = X_agric, X_ind = X_ind, X_polit = X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
fit.rgcca=rgcca(blocks = A, connection = C, 
                tau = rep(1, 3), ncomp = rep(2, 3))
                
###############               
# sample plot #
###############

# Defaut call: First component of the last block vs second component of the 
# last block
plot(fit.rgcca, type="ind", resp = status)

# horizontal axis: First component of the first block
# vertical axis: First component of the second block
plot(fit.rgcca, type="ind", block = 1:2, comp = 1, resp = status)

# horizontal axis: First component of the first block
# vertical axis: Second component of the first block
plot(fit.rgcca, type="ind", block = 1, comp = 1:2, resp = status)

######################
# Correlation circle #
######################
# with superblock
fit.mcia = rgcca(blocks=A, scheme = "factorial", ncomp = rep(2, 4), 
                 tau = c(1, 1, 1, 0), superblock = TRUE)
plot(fit.mcia, type="both", resp = status, overlap = FALSE)
                                                     
plot(fit.rgcca, type="cor")
plot(fit.rgcca, type="weight")
plot(fit.rgcca, type="ind")
plot(fit.rgcca, type="var")
plot(fit.rgcca, type="both")
plot(fit.rgcca, type="ave")
plot(fit.rgcca, type="network")
}
