% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca.R
\name{plot.rgcca}
\alias{plot.rgcca}
\title{Plot for RGCCA}
\usage{
\method{plot}{rgcca}(
  x,
  type = "weights",
  block = seq_along(x$call$blocks),
  comp = 1,
  response = as.factor(rep(1, NROW(x$Y[[1]]))),
  display_order = TRUE,
  title = NULL,
  cex = 1,
  cex_sub = 12 * cex,
  cex_main = 14 * cex,
  cex_lab = 12 * cex,
  cex_point = 3 * cex,
  n_mark = 30,
  colors = NULL,
  shapes = NULL,
  show_labels = TRUE,
  repel = FALSE,
  ...
)
}
\arguments{
\item{x}{A fitted RGCCA object (see \code{\link[RGCCA]{rgcca}})}

\item{type}{A character string: 'samples', 'weights', 'loadings',
'cor_circle', both', 'ave' (see details).}

\item{block}{A numeric corresponding to the block(s) to plot.}

\item{comp}{A numeric vector indicating the components to consider.}

\item{response}{A vector coloring the points in the "samples" plot.}

\item{display_order}{A logical value for ordering the variables.}

\item{title}{A character string giving the title of the plot.}

\item{cex}{A numeric defining the size of the objects in the plot. Default
is one.}

\item{cex_sub}{A numeric defining the font size of the subtitle. Default is
12 * cex.}

\item{cex_main}{A numeric defining the font size of the title. Default is
14 * cex.}

\item{cex_lab}{A numeric defining the font size of the labels. Default is
12 * cex.}

\item{cex_point}{A numeric defining the font size of the points. Default is
3 * cex.}

\item{n_mark}{An integer defining the maximum number of bars plotted in the
"weights" and "loadings" plots.}

\item{colors}{Colors used in the plots. Default is based on the
colorblind-friendly palette recommended in ggplot.}

\item{shapes}{Shapes used for the points in the "samples" and "cor_circle"
plots. Default is the first five shapes used in ggplot.}

\item{show_labels}{A logical value for showing the labels in plots "samples"
and "cor_circle".}

\item{repel}{A logical value for repelling text labels from each other.}

\item{...}{additional graphical parameters}
}
\value{
A ggplot2 plot object.
}
\description{
Plot different outputs of the results obtained by a rgcca function
}
\details{
\itemize{
\item "samples" for sample plot. The blocks (block argument) and components
(comp) that will be used on the horizontal and the vertical axes to plot the
individuals: (Y[[block[1]]][, comp[1]], Y[[block[2]]][,comp[2]]). Points can
be colored according to the response argument. The colors of the points can
be modified with the colors argument.
\item "weights": barplot of the block weight vector for one
specific block/component. The weights are sorted from the highest to
the lowest and only the highest are displayed. The number of displayed
weights can be set with n_marks.
\item "loadings": barplot of the block-loading vector. Variables are sorted
in decreasing correlations and only the highest
correlations are displayed. The number of displayed correlations can be set
with n_marks (defaut value = 30).
\item  "cor_circle" for correlation circle.
\item "both": displays both sample plot and correlation circle (implemented
only for one block and at least when two components are asked (ncomp >= 2).
\item "ave": displays the average variance explained for each block.}
}
\examples{
data(Russett)
status <- colnames(Russett)[9:11][apply(Russett[, 9:11], 1, which.max)]
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[, c("demostab", "dictator")])
A <- list(X_agric = X_agric, X_ind = X_ind, X_polit = X_polit)
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
fit.rgcca <- rgcca(
  blocks = A, connection = C,
  tau = rep(1, 3), ncomp = rep(2, 3)
)

###############
# sample plot #
###############
# horizontal axis: First component of the first block
# vertical axis: First component of the second block
plot(fit.rgcca, type = "samples", block = 1:2, comp = 1, response = status)


######################
# all types of plots #
######################
# with superblock
fit.mcoa <- rgcca(
  blocks = A, scheme = "factorial", ncomp = rep(2, 4),
  tau = c(1, 1, 1, 0), superblock = TRUE
)

plot(fit.mcoa, type = "both", response = status)
plot(fit.rgcca, type = "loadings")
plot(fit.rgcca, type = "weight")
plot(fit.rgcca, type = "sample")
plot(fit.rgcca, type = "cor_circle")
plot(fit.rgcca, type = "ave")
}
