% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rgcca.analyze.R
\name{plot.rgcca}
\alias{plot.rgcca}
\title{Regularized Generalized Canonical Correlation Analysis (RGCCA)}
\usage{
\method{plot}{rgcca}(resRgcca, i_block, ...)
}
\arguments{
\item{resRgcca}{Result of rgcca function}

\item{i_block}{number of block to plot}

\item{...}{parameters from plot_ind or plot_var_2D}
}
\description{
plot 
Plots
}
\examples{
data(Russett)
X_agric =as.matrix(Russett[,c("gini","farm","rent")]);
X_ind = as.matrix(Russett[,c("gnpr","labo")]);
X_polit = as.matrix(Russett[ , c("demostab", "dictator")]);
A = list(X_agric, X_ind, X_polit);
C = matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3);
resRgcca=rgcca(blocks=A,connection=C,
tau=rep(1,3),ncomp=rep(2,3),superblock=FALSE)
}
