% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Plots RGCCA components in a bi-dimensional space}
\usage{
plot2D(
  rgcca_res,
  df,
  title = "",
  group = 1,
  name_group = "Response",
  compx = 1,
  compy = 2,
  i_block = 1,
  p = NULL,
  text = TRUE,
  i_block_y = i_block,
  colors = NULL,
  collapse = FALSE,
  no_overlap = FALSE,
  cex = 1,
  cex_main = 25 * cex,
  cex_sub = 16 * cex,
  cex_point = 3 * cex,
  cex_lab = 19 * cex
)
}
\arguments{
\item{rgcca_res}{Result of rgcca function}

\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{group}{A vector of character with levels used to color the points}

\item{name_group}{A character giving the type of groups (either "Blocs" or}

\item{compx}{Integer corresponding to the number of x-component}

\item{compy}{Integer corresponding to the number of y-component}

\item{i_block}{Integer corresponding to the first block}

\item{p}{A ggplot object}

\item{text}{If TRUE, labels are plotted. If FALSE points are plotted.}

\item{i_block_y}{Integer corresponding to the second block}

\item{colors}{A vectof of character to color quantitative data}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{no_overlap}{If TRUE, the potential overlaps are reduced}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Plots RGCCA components in a bi-dimensional space
}
