% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Plot RGCCA components in a bi-dimensional space}
\usage{
plot2D(
  rgcca_res,
  df,
  title = "",
  group = 1,
  name_group = "Response",
  compx = 1,
  compy = 2,
  i_block = 1,
  p = NULL,
  text = TRUE,
  i_block_y = i_block,
  colors = NULL,
  collapse = FALSE,
  no_overlap = FALSE,
  cex = 1,
  cex_main = 14 * cex,
  cex_sub = 12 * cex,
  cex_point = 3 * cex,
  cex_lab = 10 * cex
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{df}{A data.frame}

\item{title}{A character giving the title of the plot}

\item{group}{A vector of character with levels used to color the points}

\item{name_group}{A character giving the type of groups (either "Blocs" or}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{p}{A ggplot object}

\item{text}{A bolean to represent the points with their row names (TRUE)
or with circles (FALSE)}

\item{i_block_y}{An integer giving the index of a list of blocks (another
one, different from the one used in i_block)}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Plot RGCCA components in a bi-dimensional space
}
