% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ave.R
\name{plot_ave}
\alias{plot_ave}
\title{Histogram of Average Variance Explained}
\usage{
plot_ave(
  rgcca_res,
  cex = 1,
  title = "Average Variance Explained",
  colors = NULL,
  ...
)
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{cex}{An integer for the size of the plot parameters}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{colors}{reoresenting a vector of colors}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
Histogram of the model quality (based on Average Variance Explained)
for each blocks and sorted in decreasing order
}
\examples{
random_val = function(y=1) lapply(seq(4),
function(x) matrix(runif(4), y, 2))
rgcca_out = list(AVE = list(AVE_X = random_val()),
     a = random_val(2), 
     call = list(type = "rgcca", ncomp = rep(2, 4)))
names(rgcca_out$a) <- LETTERS[seq(4)]
class(rgcca_out) = "rgcca"
library("ggplot2")
for(i in seq(1,4))
names(rgcca_out$AVE$AVE_X[[i]]) <- c(1,2)
plot_ave(rgcca_out)
}
\seealso{
\code{\link[RGCCA]{rgccad}}, \code{\link[RGCCA]{sgcca}}
}
