% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_1D.R
\name{plot_bootstrap_1D}
\alias{plot_bootstrap_1D}
\title{Plot a fitted bootstrap object in 1D}
\usage{
plot_bootstrap_1D(
  b = NULL,
  type = "weight",
  empirical = TRUE,
  x = "estimate",
  y = "sign",
  n_mark = 50,
  display_order = TRUE,
  title = NULL,
  colors = NULL,
  comp = 1,
  display_bar = TRUE,
  i_block = length(b$bootstrap[[1]][[1]]),
  ...
)
}
\arguments{
\item{b}{A fitted bootstrap object \code{\link[RGCCA]{bootstrap}}}

\item{type}{Character string indicating the bootstrapped object to print:
block-weight vectors ("weight", default) or block-loading vectors
("loadings").}

\item{empirical}{A logical value indicating if the bootstrap confidence
intervals and p-values are derived from the empirical distribution.
(defaut: TRUE)}

\item{x}{indicator used in the plot (see details).}

\item{y}{A character string indicating for the index to color the bars
(see details).}

\item{n_mark}{The number of top variables to display in the correlation
circle}

\item{display_order}{A logical value for ordering the variables}

\item{title}{A character giving the title of the plot}

\item{colors}{Either a vector of integers (each integer corresponding to a
color) or of characters corresponding to names of colors (as "blue",see
colors()) or RGB code ("#FFFFFF").}

\item{comp}{An integer indicating the block-weight/loading vector to plot.}

\item{display_bar}{A logical value. If TRUE colorbar for significant
variables is displayed.}

\item{i_block}{An integer giving the index of the block}

\item{...}{Other parameters (see plot_histogram)}
}
\description{
Display bootstrap confidence intervals for RGCCA or the number of non-zero
occurrences for SGCCA. The bars are colored according to the
significancy  block weight vectors ('*' or 'ns'; see 'pval' in details for
\code{\link[RGCCA]{get_bootstrap}}) for RGCCA and according to the
occurrences of non-zero weights for SGCCA. In SGCCA, the significant
variables are those above the three bars, respectively, with an alpha = 0.05
(dark red), 0.01 (red) and 0.001 (light red).
}
\details{
\itemize{
\item 'estimate' of the block weight vectors
\item 'bootstrap_ratio' of the block weight/loading vectors
\item 'sign' for significant 95% bootstrap interval
\item 'occurrences' number of for non-zero occurrences
\item 'mean'  mean of the bootstraped block weight vectors
}
}
