% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_1D.R
\name{plot_bootstrap_1D}
\alias{plot_bootstrap_1D}
\title{Plot a bootstrap in 1D}
\usage{
plot_bootstrap_1D(
  b = NULL,
  df_b = NULL,
  x = "estimate",
  y = "occurrences",
  n_mark = 50,
  title = NULL,
  colors = NULL,
  comp = 1,
  bars = "sd",
  i_block = length(b$bootstrap[[1]]),
  collapse = FALSE,
  n_cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{b}{A matrix of boostrap}

\item{df_b}{Result of get_bootstrap functions or dataframe #TODO}

\item{x}{A character for the column used in the plot}

\item{y}{A character for the column to color the bars}

\item{n_mark}{An integer giving the number maximum of top variables}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{colors}{reoresenting a vector of colors}

\item{comp}{An integer giving the index of the analysis components}

\item{bars}{A character among "sd" for standard deviations, "stderr" for the standard error, "ci" for confidence interval of scores and "cim" for the confidence intervall of the mean.}

\item{i_block}{An integer giving the index of a list of blocks}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{...}{Other parameters (see plot_histogram)}
}
\description{
Histogram of the best variables of an RGCCA bootstrap with, on the x-axis,
the number of non-zero occurrences (SGCCA) or the bootstrap-ratio
(mean/sd; RCCA). The bars are colored according to the average weight of
the boostrap  (according to an ascending gradient from red to blue)
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, sparsity = 0.75, type = "sgcca")
boot = bootstrap(rgcca_out, 2, n_cores = 1)
plot_bootstrap_1D(boot, n_cores = 1)
rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(boot, n_cores = 1)
plot_bootstrap_1D(boot, n_cores = 1)
plot_bootstrap_1D(df_b = selected.var)
}
