% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_2D.R
\name{plot_bootstrap_2D}
\alias{plot_bootstrap_2D}
\title{Plot a bootstrap in 2D}
\usage{
plot_bootstrap_2D(b, x = "br", y = "occ", cex = 1, cex_sub = 16 *
  cex, cex_point = 3 * cex, cex_lab = 19 * cex)
}
\arguments{
\item{b}{A matrix of boostrap}

\item{x}{A character for the column to plot in x-axis}

\item{y}{A character for the column to plot in y-axis}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Biplot of the top variables from a SGCCA bootstrap with the number of
non-zero occurences in x-axis and the boot-ratio (mean/sd) in y-axis.
Negative weights are colored in red and the positive ones are in green.
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, tau = 0.75, type = "sgcca")
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(rgcca_out, boot, n_cores = 1)
plot_bootstrap_2D(selected.var)
rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(rgcca_out, boot, n_cores = 1)
plot_bootstrap_2D(selected.var)
}
