% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_2D.R
\name{plot_bootstrap_2D}
\alias{plot_bootstrap_2D}
\title{Plot a bootstrap in 2D}
\usage{
plot_bootstrap_2D(
  b = NULL,
  df_b = NULL,
  x = "bootstrap_ratio",
  y = "occurrences",
  title = paste("Variable selection \\nby", attributes(b)$n_boot, "bootstraps"),
  colors = NULL,
  cex = 1,
  cex_main = 14 * cex,
  cex_sub = 12 * cex,
  cex_point = 3 * cex,
  cex_lab = 10 * cex,
  comp = 1,
  i_block = length(b$bootstrap[[1]]),
  collapse = FALSE,
  n_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{b}{A matrix of boostrap}

\item{df_b}{Result of get_bootstrap functions or dataframe #TODO}

\item{x}{A character for the column to plot in x-axis}

\item{y}{A character for the column to plot in y-axis}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{colors}{reoresenting a vector of colors}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}

\item{comp}{An integer giving the index of the analysis components}

\item{i_block}{Integer corresponding to the first block}

\item{collapse}{A boolean to combine the variables of each block as result}

\item{n_cores}{An integer for the number of cores used in parallelization}
}
\description{
Biplot of the top variables from a SGCCA bootstrap with the number of
non-zero occurences in x-axis and the boot-ratio (mean/sd) in y-axis.
Negative weights are colored in red and the positive ones are in green.
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, sparsity = 0.75, type = "sgcca")
boot = bootstrap(rgcca_out, 2, n_cores = 1)
plot_bootstrap_2D(boot, n_cores = 1)
rgcca_out = rgcca(blocks)
boot = bootstrap(rgcca_out, 2, n_cores = 1)
selected.var = get_bootstrap(boot, n_cores = 1,display_order=TRUE)
print("i")
# plot_bootstrap_2D(boot, n_cores = 1)
print("j")
#plot_bootstrap_2D(df_b = selected.var,n_cores=1)
}
\seealso{
\code{\link[RGCCA]{bootstrap}}, \code{\link[RGCCA]{get_bootstrap}}
}
