% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram settings}
\usage{
plot_histogram(
  p,
  df,
  title = "",
  group = NA,
  colors = NULL,
  cex = 1,
  cex_main = 25 * cex,
  cex_sub = 16 * cex,
  cex_axis = 10 * cex
)
}
\arguments{
\item{p}{A ggplot object}

\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{group}{A vector of character giving the group for the rows}

\item{colors}{reoresenting a vector of colors}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_axis}{An integer for the size of the axis text}
}
\description{
Default font for a vertical barplot.
}
