% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram settings}
\usage{
plot_histogram(
  p,
  df,
  title = "",
  group = NA,
  colors = NULL,
  cex = 1,
  cex_main = 14 * cex,
  cex_sub = 12 * cex,
  cex_axis = 10 * cex
)
}
\arguments{
\item{p}{A ggplot object}

\item{df}{A data.frame}

\item{title}{A character giving the title of the plot}

\item{group}{A vector of character giving the group for the rows}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_main}{An integer for the size of the title}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_axis}{An integer for the size of the axis text}
}
\description{
Default font for a vertical barplot.
}
