% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot the connection between blocks}
\usage{
plot_network(
  rgcca_res,
  title = paste0("Common rows between blocks : ", NROW(rgcca_res$call$blocks[[1]])),
  cex_main = 2,
  cex_point = 23,
  colors = c("#eee685", "gray60")
)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{title}{A character giving the title of the plot}

\item{cex_main}{An integer for the size of the title}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{colors}{Either a vector of integers (each integer corresponding to a color) or of characters corresponding to names of colors (as "blue",see colors()) or RGB code ("#FFFFFF").}
}
\value{
A dataframe with tuples of connected blocks
}
\description{
Plot the connection between blocks
}
\examples{
library(igraph)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
plot_network(rgcca_out)
}
