% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permut_2D.R
\name{plot_permut_2D}
\alias{plot_permut_2D}
\title{Plot permuation in 2D}
\usage{
plot_permut_2D(perm, type = "zstat", cex = 1, cex_sub = 16 * cex,
  cex_point = 3 * cex, cex_lab = 19 * cex)
}
\arguments{
\item{perm}{A permutation object from a RGCCA analyse}

\item{type}{An string giving the type of the index to look at (among 'crit' for
the RGCCA criterion and 'zstat' for the pseudo Z-score)}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Plot permuation in 2D
}
\examples{
data("Russett")
A = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
perm <- rgcca_permutation(A, nperm = 2, n_cores = 1)
plot_permut_2D(perm)
perm <- rgcca_permutation(A, p_c1 = TRUE, nperm = 2, n_cores = 1)
plot_permut_2D(perm)
}
