% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permut_2D.R
\name{plot_permut_2D}
\alias{plot_permut_2D}
\title{Plot permuation in 2D}
\usage{
plot_permut_2D(perm, type = "zstat", cex = 1, title = NULL,
  cex_main = 14 * cex, cex_sub = 12 * cex, cex_point = 3 * cex,
  cex_lab = 10 * cex, bars = "points")
}
\arguments{
\item{perm}{A permutation object from a RGCCA analyse}

\item{type}{An string giving the type of the index to look at (among 'crit' for
the RGCCA criterion and 'zstat' for the pseudo Z-score)}

\item{cex}{general size of the text}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}

\item{cex_main}{= 25 * cex, size of the main text (title)}

\item{cex_sub}{= 16 * cex, size of the subtitle text}

\item{cex_point}{= 3 * cex, size of the point}

\item{cex_lab}{= 19 * cex, size of the labels}

\item{bars}{Among "points", "stderr" or "sd": representation of the variability}
}
\description{
Plot permuation in 2D
}
