% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_1D.R
\name{plot_var_1D}
\alias{plot_var_1D}
\title{Histogram of a fingerprint}
\usage{
plot_var_1D(rgcca, comp = 1, n_mark = 100, i_block = length(rgcca$a),
  type = "cor", collapse = FALSE, cex = 1, cex_sub = 16 * cex,
  cex_axis = 10 * cex)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{comp}{An integer giving the index of the analysis components}

\item{n_mark}{An integer giving the number of top variables to select}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A string giving the criterion to selects biomarkers : either 
"cor" for correlation between the component and the block
or "weight" for the weight of the RGCCA}

\item{collapse}{A boolean to combine the variables of each blocks as result}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_axis}{An integer for the size of the axis text}
}
\description{
Histogram of the higher outer weight vectors for a component of a block 
(by default, the superblock or the last one) analysed by R/SGCCA
}
\examples{
weights = lapply(seq(3), function(x) matrix(runif(7*2), 7, 2))
for (i in seq(3))
row.names(weights[[i]]) <- paste0(letters[i],
     letters[seq(NROW(weights[[i]]))])
weights[[4]] = Reduce(rbind, weights)
rgcca_out = list(a = weights)
names(rgcca_out$a) = LETTERS[seq(4)]
rgcca_out$blocks = lapply(rgcca_out$a, t)
rgcca_out$superblock = TRUE
# With the 1rst component of the superblock
plot_var_1D(rgcca_out, 1, type = "weight")
# With the 2nd component of the 1rst block by selecting the ten higher weights
plot_var_1D(rgcca_out, 2, 10, 1, type = "weight")
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca.analyze(blocks)
plot_var_1D(rgcca_out, collapse = TRUE)
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
