% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_1D.R
\name{plot_var_1D}
\alias{plot_var_1D}
\title{Barplot of a fingerprint}
\usage{
plot_var_1D(rgcca, comp = 1, n_mark = 100, i_block = length(rgcca$a),
  type = "cor", collapse = FALSE, title = NULL, colors = NULL, ...)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{comp}{An integer giving the index of the analysis components}

\item{n_mark}{An integer giving the number of top variables to select}

\item{i_block}{An integer giving the index of a list of blocks}

\item{type}{A string giving the criterion to selects biomarkers : either 
"cor" for correlation between the component and the block
or "weight" for the weight of the RGCCA}

\item{collapse}{A boolean to combine the variables of each blocks as result}

\item{title}{A character with the name of the space (either "Variables" or
"Samples")}
}
\description{
Barplot of the higher outer weight vectors for a component of a block 
(by default, the superblock or the last one) analysed by R/SGCCA
}
\examples{
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
plot_var_1D(rgcca_out, collapse = TRUE)
}
\seealso{
\code{\link[RGCCA]{rgccad}}, \code{\link[RGCCA]{sgcca}}
}
