% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_2D.R
\name{plot_var_2D}
\alias{plot_var_2D}
\title{Plot of variables space}
\usage{
plot_var_2D(rgcca, compx = 1, compy = 2, i_block = length(rgcca$a),
  text = TRUE, remove_var = TRUE, n_mark = 100, collapse = FALSE,
  no_overlap = FALSE, cex = 1, cex_sub = 16 * cex, cex_point = 3 *
  cex, cex_lab = 19 * cex)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{compx}{An integer giving the index of the analysis component used
for the x-axis}

\item{compy}{An integer giving the index of the analysis component used
for the y-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{text}{A bolean to represent the points with their row names (TRUE)
or with circles (FALSE)}

\item{remove_var}{A bolean to keep only the 100 variables of each
component with the biggest correlation#'}

\item{n_mark}{An integer giving the number of top variables to select}

\item{collapse}{A boolean to combine the variables of each blocks as result}

\item{no_overlap}{A boolean to avoid overlap in plotted text}

\item{cex}{An integer for the size of the plot parameters}

\item{cex_sub}{An integer for the size of the subtitle}

\item{cex_point}{An integer for the size of the points or the text in the plot}

\item{cex_lab}{An integer for the size of the axis titles}
}
\description{
Correlation circle highlighting the contribution of each variables to the
construction of the RGCCA components
}
\examples{
setMatrix = function(nrow, ncol, iter = 3) 
 lapply(
     seq(iter),
     function(x) {
         y <- matrix(runif(nrow * ncol), nrow, ncol)
         rownames(y) <- paste0("S",1:nrow)
         return(y)
     })
blocks = setMatrix(10, 5)
blocks[[4]] = Reduce(cbind, blocks)
for (i in seq(4)) {
 colnames(blocks[[i]]) = paste0( LETTERS[i],
 as.character(seq(NCOL(blocks[[i]]))))
}
coord = setMatrix(10, 2, 4)
a = setMatrix(5, 2)
a[[4]] = setMatrix(15, 2, 1)[[1]]
AVE_X = lapply(seq(4), function(x) runif(2))
rgcca_out = list(Y = coord, a = a, AVE = list(AVE_X = AVE_X), call = list(blocks = blocks))
names(rgcca_out$a) <- LETTERS[seq(4)] -> names(rgcca_out$call$blocks)
rgcca_out$call$type="rgcca"
# Using a superblock
rgcca_out$call$superblock = TRUE
rgcca_out$call$ncomp = rep(3, 4)
plot_var_2D(rgcca_out, 1, 2)
# Using the first block
plot_var_2D(rgcca_out, 1, 2, 1)
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
 politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks)
# Without superblock but with the of all variables to the first block
}
