% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAVE}
\alias{plotAVE}
\title{Histogram of Average Variance Explained}
\usage{
plotAVE(rgcca, comp = 1)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{comp}{An integer giving the index of the analysis component}
}
\description{
Histogram of the model quality (base on Average Variance Explained) for each blocks and sort in decreasing order
}
\examples{
random_val = function() lapply(1:4, function(x) runif(1))
rgcca.res = list(AVE = list(AVE_X = random_val()), a = random_val())
names(rgcca.res$a) = LETTERS[1:4]
plotAVE(rgcca.res, 1)
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
