% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFingerprint}
\alias{plotFingerprint}
\title{Histogram of a fingerprint}
\usage{
plotFingerprint(rgcca, i_comp = 1, superblock = T, n_mark = 100,
  i_block = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{i_comp}{An integer giving the index of the analysis component}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of a superblock}

\item{n_mark}{An integer giving the number of best fingerprint to select}

\item{i_block}{An integer giving the index of a list of blocks}
}
\description{
Histogram of the higher outer weight vectors for a component of a block (by default, the superblock or the last one) analysed by R/SGCCA
}
\examples{
weights = lapply(1:3, function(x) matrix(runif(10*2), 10, 2))
weights[[4]] = Reduce(rbind, weights)
rgcca.res = list(a = weights)
names(rgcca.res$a) = LETTERS[1:4]
# With the 1rst component of the superblock
plotFingerprint(rgcca.res, 1, TRUE)
# With the 2nd component of the 1rst block by selecting the 5 higher weights
plotFingerprint(rgcca.res, 2, FALSE, 10, 1)
}
\seealso{
\code{\link[RGCCA]{rgcca}}, \code{\link[RGCCA]{sgcca}}
}
