% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Histogram settings}
\usage{
plotHistogram(p, df, title = "", color = "black", low_col = "khaki2",
  high_col = "coral3")
}
\arguments{
\item{p}{A ggplot object.}

\item{df}{A dataframe with a column named "order"}

\item{title}{A character string giving a graphic title}

\item{color}{A vector of character giving the colors for the rows}

\item{low_col}{A character giving the color used for the lowest part of the gradient}

\item{high_col}{A character giving the color used for the highest part of the gradient}
}
\description{
Default font for a vertical barplot.
}
\examples{
df = data.frame(x = runif(30), order = 30:1)
library("ggplot2")
p = ggplot(df, aes(order, x))
plotHistogram(p, df, "This is my title")
# Add colors per levels of a variable
df$color = rep(c(1,2,3), each=10)
p = ggplot(df, aes(order, x, fill = color))
plotHistogram(p, df, "Histogram", as.character(df$color))
}
