% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSamplesSpace}
\alias{plotSamplesSpace}
\title{Plot of samples space}
\usage{
plotSamplesSpace(rgcca, response, compX = 1, compY = 2,
  i_block = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{response}{A vector of characters corresponding either to a qualitative variable with levels or a continuous variable}

\item{compX}{An integer giving the index of the analysis component used for the x-axis}

\item{compY}{An integer giving the index of the analysis component used for the y-axis}

\item{i_block}{An integer giving the index of a list of blocks}
}
\description{
Plots samples on two components of a RGCCA
}
\examples{
coord = lapply(1:3, function(x) matrix(runif(15 * 2, min = -1), 15, 2))
AVE_X = lapply(1:3, function(x) runif(2))
rgcca.res = list(Y = coord, AVE = list(AVE_X = AVE_X))
# Using a superblock
plotSamplesSpace(rgcca.res, rep(LETTERS[1:3], each = 5))
# Using the first block
plotSamplesSpace(rgcca.res, runif(15, min=-15, max = 15), 1, 2, 1)
}
