% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpace}
\alias{plotSpace}
\title{Plot of components space}
\usage{
plotSpace(rgcca, df, title, group, name_group, comp_x = 1, comp_y = 2,
  i_block = 1, p = NULL)
}
\arguments{
\item{rgcca}{A list giving the results of a R/SGCCA}

\item{df}{A dataframe}

\item{title}{A character with the name of the space (either "Variables" or "Samples")}

\item{group}{A vector of character with levels used to color the points}

\item{name_group}{A character giving the type of groups (either "Blocs"  or "Response")}

\item{comp_x}{An integer giving the index of the analysis component used for the x-axis}

\item{comp_y}{An integer giving the index of the analysis component used for the y-axis}

\item{i_block}{An integer giving the index of a list of blocks}

\item{p}{A ggplot object}
}
\description{
Plots RGCCA components in a bi-dimensional space
}
\examples{
df = as.data.frame(matrix(runif(20*2, min = -1), 20, 2))
AVE =  lapply(1:4, function(x) runif(2))
rgcca.res = list(AVE = list(AVE_X = AVE))
plotSpace(rgcca.res, df, "Samples", rep(c("a","b"), each=10), "Response")
}
