% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootstrap.r
\name{print.bootstrap}
\alias{print.bootstrap}
\title{Print bootstrap}
\usage{
\method{print}{bootstrap}(x, ...)
}
\arguments{
\item{x}{A bootstrap object (see \code{\link[RGCCA]{bootstrap}} )}

\item{...}{Further arguments in print}
}
\value{
A matrix containing for each variables of each blocks, the means, 95\% intervals, bootstrap ratio, p-values and other statistics (see details)
}
\description{
Print bootstrap
}
\details{
\itemize{
\item 'estimate' for RGCCA weights
\item 'mean' for the mean of the bootstrap weights
\item 'sd' for the standard error of the bootstrap weights
\item 'lower/upper_band' for the lower and upper intervals from to the 'bar' parameter
\item 'p.vals' for p-values. The occurrences of  are distributed in accordance with the law
\item 'BH' for Benjamini-Hochberg p-value adjustments
}
}
