% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootstrap.r
\name{print.bootstrap}
\alias{print.bootstrap}
\title{Print bootstrap}
\usage{
\method{print}{bootstrap}(x, ...)
}
\arguments{
\item{x}{results of a bootstrap function}

\item{...}{Further arguments in print}
}
\description{
Print bootstrap
}
