% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bootstrap.r
\name{print.bootstrap}
\alias{print.bootstrap}
\title{Print bootstrap}
\usage{
\method{print}{bootstrap}(x, type = "weight", empirical = TRUE, display_order = FALSE, ...)
}
\arguments{
\item{x}{A fitted bootstrap object (see \code{\link[RGCCA]{bootstrap}})}

\item{type}{Character string indicating the bootstrapped object to print:
block-weight vectors ("weight", default) or block-loading vectors
("loadings").}

\item{empirical}{A logical value indicating if the bootstrap confidence
intervals and p-values are derived from the empirical distribution.
(defaut: TRUE)}

\item{display_order}{A logical value for ordering the variables}

\item{...}{Further arguments in print
the means, 95\% intervals, bootstrap ratio, p-values and other statistics.}
}
\value{
A matrix containing for each variable:
\itemize{
\item 'estimate' for block weight or loading vectors.
\item 'mean' for the mean of the bootstrap block weight/loading vectors.
\item 'sd' for the boostrapped estimate of the standard deviation of the
block weight/loading vectors.
\item 'lower/upper_bound' for the lower and upper intervals.
\item 'pvals' for p-values. In the case of SGCCA, the occurrences of the
bootstrap weights are distributed according to the binomial distribution
while for RGCCA, their absolute value divided by their standard deviation
follows a normal distribution.
\item 'adjust.pval' for ajusted p-value (fdr correction by default)
}
}
\description{
Print a bootstrap object
}
