% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cval.r
\name{print.cval}
\alias{print.cval}
\title{print.cval}
\usage{
\method{print}{cval}(x, type = "sd", ...)
}
\arguments{
\item{x}{A rgcca_cv object (see \link{rgcca_cv})}

\item{type}{Character string indicating the statistics in the box plots:
mean plus or minus standard deviation ("sd", default),
\itemize{
\item "sd" (default): the middle bar corresponds to the mean and limits of
the boxes are given by the mean plus or minus the standard deviation.
\item "stderr": the middle bar corresponds to the mean and limits of
the boxes are given by the mean plus or minus the standard deviation divided
by the square roots of the number of folds.
\item "quantile": the middle bar corresponds to the median and limits of
the boxes are given by the 25% and 75% quantiles.
\item "points": box plots are removed and only the points are kept.}}

\item{...}{Further print options}
}
\description{
print.cval
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
res <- rgcca_cv(blocks,
  response = 3, method = "rgcca", par_type = "tau",
  par_value = c(0, 0.2, 0.3), n_run = 1, n_cores = 1
)
print(res)
}
