% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cval.r
\name{print.cval}
\alias{print.cval}
\title{print.cval}
\usage{
\method{print}{cval}(x, bars = "quantile", ...)
}
\arguments{
\item{x}{A rgcca_cv object (see \link{rgcca_cv})}

\item{bars}{standard deviation (sd) of the cross-validated MSE}

\item{...}{Further print options}
}
\description{
print.cval
}
\examples{
data("Russett")
blocks <- list(
  agriculture = Russett[, seq(3)],
  industry = Russett[, 4:5],
  politic = Russett[, 6:11]
)
res <- rgcca_cv(blocks,
  response = 3, method = "rgcca", par_type = "tau",
  par_value = c(0, 0.2, 0.3), n_run = 1, n_cores = 1
)
print(res)
}
