% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.perm.R
\name{print.permutation}
\alias{print.permutation}
\title{Print the call of rgcca_permutation results}
\usage{
\method{print}{permutation}(x, ...)
}
\arguments{
\item{x}{A permutation object (see \code{\link{rgcca_permutation}})}

\item{...}{other parameters used in print (for the displaying of matrices)}
}
\description{
Print the call of rgcca_permutation results
}
\examples{
data("Russett")
A <- list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
perm <- rgcca_permutation(A, n_run = 2, n_cores = 1)
print(perm)
}
