% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sgcca.permute.R
\name{print.permutation}
\alias{print.permutation}
\title{Prints the results of permutation rgcca}
\usage{
\method{print}{permutation}(x, ...)
}
\arguments{
\item{x}{A rgcca_permutation object (see  \code{\link[RGCCA]{rgcca_permutation}} )}

\item{...}{Further print parameters}
}
\description{
Prints the results of permutation rgcca
}
\examples{
data('Russett')
A = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
res = rgcca_permutation(A, nperm = 5, n_cores = 1)
print(res)
}
