% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sgcca.permute.R
\name{print.permutation}
\alias{print.permutation}
\title{Prints the results of permutation rgcca}
\usage{
\method{print}{permutation}(x)
}
\arguments{
\item{x}{result of rgcca_permutation}
}
\description{
Prints the results of permutation rgcca
}
