% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_comp.R
\name{print_comp}
\alias{print_comp}
\title{Print the variance of a component}
\usage{
print_comp(rgcca_res, n = 1, i = length(rgcca_res$AVE$AVE_X), outer = FALSE)
}
\arguments{
\item{rgcca_res}{A fitted RGCCA object (see  \code{\link[RGCCA]{rgcca}})}

\item{n}{An integer giving the index of the analysis component}

\item{i}{An integer giving the index of a list of blocks}

\item{outer}{A boolean for ave plot case}
}
\value{
A character for the variance on the component
}
\description{
Print the percent of explained variance for a component of a block
(by default, the superblock or the last one) analysed by R/SGCCA
}
\seealso{
\code{\link[RGCCA]{rgccad}}, \code{\link[RGCCA]{sgcca}}
}
