% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_null_sd.R
\name{remove_null_sd}
\alias{remove_null_sd}
\title{Remove columns having a 0 standard deviation}
\usage{
remove_null_sd(list_m, column_sd_null = NULL)
}
\arguments{
\item{list_m}{A list of dataframe}

\item{column_sd_null}{Either NULL or a list of named vectors. If NULL, the
function will search for variables with null variance in each block. If not
NULL, this list defines for each block the index of the variables that are
of null variance (see the 'Value' section for more details about the content
of this list). In both cases, these variables are removed.}
}
\value{
\item{list_m}{A list of dataframe.}

\item{column_sd_null}{Either NULL, if not a single variable was
removed, or a list of the same size as the number of blocks. In the last
situation, each element of this list is again NULL if not a single variable
was removed from the current block, or a named vector indicating the former
index of the removed variables along with their name.}
}
\description{
Remove columns having a 0 standard deviation
}
\keyword{internal}
