% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcca.analyze.R
\name{rgcca.analyze}
\alias{rgcca.analyze}
\title{Performs a r/sgcca}
\usage{
rgcca.analyze(blocks, connection = 1 - diag(length(blocks)),
  response = NULL, superblock = TRUE, tau = rep(1, length(blocks)),
  ncomp = rep(2, length(blocks)), type = "rgcca", verbose = TRUE,
  scheme = "factorial", scale = TRUE, init = "svd", bias = TRUE,
  tol = 1e-08)
}
\arguments{
\item{blocks}{A list of matrix}

\item{connection}{A matrix giving the connection between the blocks}

\item{response}{An integer giving the index of a block considered as a 
response among a list of blocks}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{tau}{A vector of float (or character for 'optimal' setting) giving the
shrinkage parameter for covariance maximization}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{type}{A character giving the type of analysis}

\item{verbose}{A boolean to display the progress of the analysis}

\item{scheme}{A character giving the link function for covariance maximization}

\item{scale}{A boolean scaling the blocks}

\item{init}{A character among "svd" (Singular Value Decompostion) or "random"
for alorithm initialization}

\item{bias}{A boolean for a biased variance estimator}

\item{tol}{An integer for the stopping value for convergence}
}
\value{
A RGCCA object
}
\description{
Performs a r/sgcca with predefined parameters
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca.analyze(blocks)
}
