% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcca.R
\name{rgcca}
\alias{rgcca}
\title{Performs a r/sgcca}
\usage{
rgcca(blocks, connection = 1 - diag(length(blocks)), response = NULL,
  superblock = TRUE, tau = rep(1, length(blocks)), sparsity = rep(1,
  length(blocks)), ncomp = rep(2, length(blocks)), type = "rgcca",
  verbose = FALSE, scheme = "factorial", scale = TRUE,
  init = "svd", bias = TRUE, tol = 1e-08, quiet = TRUE,
  sameBlockWeight = TRUE, method = "complete", knn.k = "all",
  knn.output = "weightedMean", knn.klim = NULL,
  knn.sameBlockWeight = TRUE, pca.ncp = 1)
}
\arguments{
\item{blocks}{A list of matrix}

\item{connection}{A matrix giving the connection between the blocks}

\item{response}{An integer giving the index of a block considered as a 
response among a list of blocks}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{tau}{tau is either a \eqn{1 \times J} vector or a \eqn{\mathrm{max}(ncomp) \times J} matrix, and contains the values 
of the shrinkage parameters (default: tau = 1, for each block and each dimension).
If tau = "optimal" the shrinkage paramaters are estimated for each block and each dimension using the Schafer and Strimmer (2005)
analytical formula . If tau is a \eqn{1\times J} numeric vector, tau[j] is identical across the dimensions of block \eqn{\mathbf{X}_j}. 
If tau is a matrix, tau[k, j] is associated with \eqn{\mathbf{X}_{jk}} (\eqn{k}th residual matrix for block \eqn{j})}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{type}{A character giving the type of analysis: c('rgcca', 'cpca-w', 'gcca', 'hpca', 'maxbet-b', 'maxbet', 'maxdiff-b','maxdiff', 'maxvar-a', 'maxvar-b', 'maxvar', 'niles', 'r-maxvar', 'rcon-pca',
'ridge-gca', 'sabscor', 'ssqcor', 'ssqcor', 'ssqcov-1', 'ssqcov-2', 'ssqcov',
'sum-pca', 'sumcor', 'sumcov-1', 'sumcov-2', 'sumcov.', 'sabscov', 'plspm','cca', 'ra', 'ifa', 'pls','pca')}

\item{verbose}{A boolean to display the progress of the analysis}

\item{scheme}{A character giving the link function for covariance maximization}

\item{scale}{A boolean scaling the blocks}

\item{init}{A character among "svd" (Singular Value Decompostion) or "random"
for algorithm initialization}

\item{bias}{A boolean for a biased variance estimator}

\item{tol}{An integer for the stopping value for convergence}

\item{quiet}{A boolean hidding the warnings}

\item{sameBlockWeight}{If TRUE, all blocks are weighted by their own variance: all the blocks have the same weight}

\item{method}{Either a character corresponding to the used method ("complete","knn","em","sem") or a function taking a list of J blocks (A) as only parameter and returning the imputed list.}

\item{knn.k}{Number of k nearest neighbors}

\item{knn.output}{"mean", "random" or "weightedMean" : returns respectively the average of the k nearest neigbors, one selected randomly, or an average weighted by the distance of the k NN}

\item{knn.klim}{k limits (if k is not a number, optimal k between klim[1] and klim[2] is calculated )}

\item{knn.sameBlockWeight}{if TRUE the distance for Nearest Neigbors takes the size of blocks into account}

\item{pca.ncp}{Number of components chosen in PCA}
}
\value{
A RGCCA object
}
\description{
Performs a r/sgcca with predefined parameters
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca(blocks)
}
