% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.R, R/rgcca_crossvalidation.R
\name{rgcca_crossvalidation}
\alias{rgcca_crossvalidation}
\title{Cross-validation}
\usage{
rgcca_crossvalidation(rgcca_res, i_block = length(rgcca_res$call$blocks),
  validation = "loo", type = "regression", fit = "lm",
  new_scaled = TRUE, k = 5, n_cores = parallel::detectCores() - 1)

rgcca_crossvalidation(rgcca_res, i_block = length(rgcca_res$call$blocks),
  validation = "loo", type = "regression", fit = "lm",
  new_scaled = TRUE, k = 5, n_cores = parallel::detectCores() - 1)
}
\arguments{
\item{rgcca_res}{A list giving the results of a R/SGCCA}

\item{i_block}{An integer giving the index of a list of blocks}

\item{validation}{A character given the method for validation aong test 
(for test-train sets), kfold (for k-fold with k=5 by default) and loo 
(for leave-one-out)}

\item{type}{A character corresponding to the type of prediction among : regression or classification}

\item{fit}{A character giving the function used to compare the trained and the tested models}

\item{new_scaled}{A boolean scaling the blocks to predict}

\item{k}{An integer given the parameter for k-fold method (5, by default)}

\item{n_cores}{An integer for the number of cores used in parallelization}
}
\description{
Cross-validation for RGCCA

Uses cross-validation to validate a predictive model of RGCCA
}
\examples{
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, response = 3)
rgcca_crossvalidation(rgcca_out, validation = "kfold", k = 5, n_cores = 1)
library(RGCCA)
data("Russett")
blocks = list(agriculture = Russett[, seq(3)], industry = Russett[, 4:5],
    politic = Russett[, 6:11] )
rgcca_out = rgcca(blocks, response = 3)
rgcca_crossvalidation(rgcca_out, validation = "kfold", k = 5, n_cores = 1)
rgcca_crossvalidation(rgcca_out,  validation = "test", n_cores = 1)$scores
rgcca_crossvalidation(rgcca_out, n_cores = 1)
}
\seealso{
\link{rgcca}, \link{rgcca_predict}, \link{plot.predict}
}
