% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcca_cv.r
\name{rgcca_cv}
\alias{rgcca_cv}
\title{rgcca_cv}
\usage{
rgcca_cv(blocks, type = "rgcca", response = length(blocks),
  par = "ncomp", par_value = NULL, validation = "kfold",
  type_cv = "regression", fit = "lm", k = 5, n_cv = 20,
  one_value_per_cv = FALSE, n_cores = parallel::detectCores() - 1,
  quiet = TRUE, superblock = FALSE, ...)
}
\arguments{
\item{blocks}{A list that contains the \eqn{J} blocks of variables \eqn{\mathbf{X_1}, \mathbf{X_2}, ..., \mathbf{X_J}}.}

\item{type}{A character giving the type of analysis: c('rgcca', 'cpca-w', 'gcca', 'hpca', 'maxbet-b', 'maxbet', 'maxdiff-b','maxdiff', 'maxvar-a', 'maxvar-b', 'maxvar', 'niles', 'r-maxvar', 'rcon-pca',
'ridge-gca', 'sabscor', 'ssqcor', 'ssqcor', 'ssqcov-1', 'ssqcov-2', 'ssqcov',
'sum-pca', 'sumcor', 'sumcov-1', 'sumcov-2', 'sumcov.', 'sabscov', 'plspm','cca', 'ra', 'ifa', 'pls','pca')}

\item{response}{Integer representing the number of the response block. Indeed, when a supervised analysis is chosen (plspm, pls), the response block has to be specified.}

\item{par}{"sparsity", "tau" or "ncomp".}

\item{par_value}{Grid of values to be tested. #TODO}

\item{validation}{A character given the method for validation aong test 
(for test-train sets), kfold (for k-fold with k=5 by default) and loo 
(for leave-one-out)}

\item{type_cv}{type of crossvalidation. Default to "regression" #TODO}

\item{fit}{A character giving the function used to compare the trained and the tested models}

\item{k}{An integer given the parameter for k-fold method (5, by default)}

\item{n_cv}{number of cross-validation. Default to 1 (with kfold option and 5)}

\item{one_value_per_cv}{If TRUE, the k values obtained for each k-fold steps are averaged}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{quiet}{If TRUE, does not print warnings}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{...}{other RGCCA parameters # TODO}
}
\description{
This function is dedicated to supervised approaches (with response parameter in rgcca not NULL). By default, it is the last block in blocks.
It allows to test a grid of parameters (sparsity, ncomp or tau) with cross-validation approaches. 
This cross-validation is based on RMSE for quantitative response.
}
\examples{
data("Russett")
blocks <- list(
    agriculture = Russett[, seq(3)],
    industry = Russett[, 4:5],
    politic = Russett[, 6:11])
    res=rgcca_cv(blocks, type="rgcca",par="tau",par_value=c(0,0.2,0.3),n_cv=1,n_cores=1)
    
}
