% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcca_cv.r
\name{rgcca_cv}
\alias{rgcca_cv}
\title{rgcca_cv}
\usage{
rgcca_cv(blocks, type = "rgcca", response = NULL, par = "tau",
  par_value = NULL, validation = "kfold", type_cv = "regression",
  fit = "lm", k = 5, n_cv = 1, one_value_per_cv = FALSE,
  n_cores = parallel::detectCores() - 1, quiet = TRUE,
  superblock = FALSE, scale = TRUE, scale_block = TRUE,
  tol = 1e-06, scheme = "factorial", method = "nipals",
  rgcca_res = NULL, parallelization = NULL, tau = rep(1,
  length(blocks)), ncomp = rep(1, length(blocks)), sparsity = rep(1,
  length(blocks)), init = "svd", bias = TRUE, ...)
}
\arguments{
\item{blocks}{A list that contains the \eqn{J} blocks of variables \eqn{\mathbf{X_1}, \mathbf{X_2}, ..., \mathbf{X_J}}.}

\item{type}{A character giving the type of analysis: c('rgcca', 'cpca-w', 'gcca', 'hpca', 'maxbet-b', 'maxbet', 'maxdiff-b','maxdiff', 'maxvar-a', 'maxvar-b', 'maxvar', 'niles', 'r-maxvar', 'rcon-pca',
'ridge-gca', 'sabscor', 'ssqcor', 'ssqcor', 'ssqcov-1', 'ssqcov-2', 'ssqcov',
'sum-pca', 'sumcor', 'sumcov-1', 'sumcov-2', 'sumcov.', 'sabscov', 'plspm','cca', 'ra', 'ifa', 'pls','pca')}

\item{response}{Integer representing the number of the response block. Indeed, when a supervised analysis is chosen (plspm, pls), the response block has to be specified.}

\item{par}{"sparsity", "tau" or "ncomp".}

\item{par_value}{Grid of values to be tested. #TODO}

\item{validation}{A character given the method for validation aong test 
(for test-train sets), kfold (for k-fold with k=5 by default) and loo 
(for leave-one-out)}

\item{type_cv}{type of crossvalidation. Default to "regression" #TODO}

\item{fit}{A character giving the function used to compare the trained and the tested models}

\item{k}{An integer given the parameter for k-fold method (5, by default)}

\item{n_cv}{number of cross-validation. Default to 1 (with kfold option and 5)}

\item{one_value_per_cv}{If TRUE, the k values obtained for each k-fold steps are averaged}

\item{n_cores}{An integer for the number of cores used in parallelization}

\item{quiet}{If TRUE, does not print warnings}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{scale}{If scale = TRUE, each block is standardized to zero means and unit variances (default: TRUE).}

\item{scale_block}{TRUE by default : each block have the same weight in the RGCCA analysis. If FALSE, the weight of each block depends on the number of variables of the block}

\item{tol}{The stopping value for convergence.}

\item{scheme}{The value is "horst", "factorial", "centroid" or the g function (default: "centroid").}

\item{method}{Either a character corresponding to the used method ("complete","knn","em","sem") or a function taking a list of J blocks (A) as only parameter and returning the imputed list. 
\itemize{
\item{\code{"mean"}}{ corresponds to an imputation by the colmeans}
\item{\code{"complete"}}{ corresponds to run RGCCA only on the complete subjects (subjects with missing data are removed)}
\item{\code{"nipals"}}{ corresponds to run RGCCA on all available data (NIPALS algorithm)}
\item{\code{"em"}}{ corresponds to impute the data with EM-type algorithms}
\item{\code{"sem"}}{ corresponds to impute the data with EM-type algorithms with superblock approach}
\item{\code{"knn1"}}{ corresponds to impute the data with the 1-Nearest Neighbor. 1 can be replace by another number (such as knn3) to impute with the 3-Nearest Neighbors.}}}

\item{rgcca_res}{a result of rgcca (from whom all parameters will be imported)}

\item{parallelization}{if TRUE parallelization is run, if FALSE, no parallelisation is run. If NULL (default) parallelization is always used except for Windows in case of length(nperm)<10}

\item{tau}{Used for type="rgcca" only. tau is either a \eqn{1 \times J} vector or a \eqn{\mathrm{max}(ncomp) \times J} matrix, and contains the values 
of the regularization parameters (default: tau = 1, for each block and each dimension).
If tau = "optimal" the regularization paramaters are estimated for each block and each dimension using the Schafer and Strimmer (2005)
analytical formula . If tau is a \eqn{1\times J} numeric vector, tau[j] is identical across the dimensions of block \eqn{\mathbf{X}_j}. 
If tau is a matrix, tau[k, j] is associated with \eqn{\mathbf{X}_{jk}} (\eqn{k}th residual matrix for block \eqn{j})}

\item{ncomp}{A \eqn{1 \times J} vector that contains the numbers of components for each block (default: rep(1, length(A)), which gives one component per block.). It can be estimated by using \link{rgcca_permutation}.}

\item{sparsity}{Used for type="sgcca" or "spls" only. A vector containing the sparsity coefficients (length J, between 0 and 1). It can be estimated by using \link{rgcca_permutation}.}

\item{init}{The mode of initialization to use in RGCCA algorithm. The alternatives are either by Singular Value Decompostion ("svd") or random ("random") (Default: "svd").}

\item{bias}{A logical value for biaised or unbiaised estimator of the var/cov (default: bias = TRUE).}

\item{...}{Further graphical parameters (see plot2D functions)}
}
\description{
This function is dedicated to supervised approaches (with response parameter in rgcca not NULL). By default, it is the last block in blocks.
It allows to test a grid of parameters (sparsity, ncomp or tau) with cross-validation approaches. 
This cross-validation is based on RMSE for quantitative response.
}
\examples{
data("Russett")
blocks <- list(
    agriculture = Russett[, seq(3)],
    industry = Russett[, 4:5],
    politic = Russett[, 6:11])
    res=rgcca_cv(blocks,response=3, type="rgcca",par="tau",par_value=c(0,0.2,0.3),n_cv=1,n_cores=1)
    
}
