% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_analysis.R
\name{select_analysis}
\alias{select_analysis}
\title{Define the analysis parameters}
\usage{
select_analysis(blocks = blocks, connection = 1 - diag(length(blocks)),
  tau = rep(1, length(blocks)), ncomp = rep(1, length(blocks)),
  scheme = "centroid", superblock = TRUE, type = "rgcca",
  verbose = TRUE, quiet = FALSE)
}
\arguments{
\item{blocks}{A list of matrix}

\item{connection}{A matrix giving the connection between the blocks}

\item{tau}{A vector of float (or character for 'optimal' setting) giving the
shrinkage parameter for covariance maximization}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{scheme}{A character giving the link function for covariance maximization}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{type}{A character giving the type of analysis: c('rgcca', 'cpca-w', 'gcca', 'hpca', 'maxbet-b', 'maxbet', 'maxdiff-b','maxdiff', 'maxvar-a', 'maxvar-b', 'maxvar', 'niles', 'r-maxvar', 'rcon-pca',
'ridge-gca', 'sabscor', 'ssqcor', 'ssqcor', 'ssqcov-1', 'ssqcov-2', 'ssqcov',
'sum-pca', 'sumcor', 'sumcov-1', 'sumcov-2', 'sumcov.', 'sabscov', 'plspm','cca', 'ra', 'ifa', 'pls','pca')}

\item{verbose}{A boolean displaying the warnings}

\item{quiet}{A boolean hidding the warnings}
}
\value{
\item{blocks}{A list of matrix}

\item{scheme}{A character giving the link function for covariance
maximization}

\item{tau}{A vector of float (or character for 'optimal' setting) giving
the shrinkage parameter for covariance maximization}

\item{ncomp}{A vector of integer giving the number of component for each
blocks}

\item{connection}{matrix giving the connection between the blocks}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE)
of a superblock}
}
\description{
Define the correct parameters according to the type of the analysis
}
