% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_analysis.R
\name{select_analysis}
\alias{select_analysis}
\title{Define the parameters associated with each well-known multi-block component
method}
\usage{
select_analysis(
  blocks,
  connection = 1 - diag(length(blocks)),
  penalty = rep(1, length(blocks)),
  ncomp = rep(1, length(blocks)),
  scheme = "centroid",
  superblock = TRUE,
  type = "rgcca",
  verbose = TRUE,
  quiet = FALSE,
  response = NULL
)
}
\arguments{
\item{blocks}{A list of blocks}

\item{connection}{A symmetric matrix (J*J) that describes the relationships 
between blocks. Two values are accepted : '1' for a connection between two 
blocks, or '0' otherwise.}

\item{penalty}{A vector of 1*J floats (or character for 'optimal' setting) 
giving theshrinkage parameter for covariance maximization}

\item{ncomp}{A vector of 1*J integers giving the number of component for 
each blocks}

\item{scheme}{A character or a function giving the link function for 
covariance maximization among "horst" (the identity function), "factorial"
 (the squared values), "centroid" (the absolute values). Only, the horst 
 scheme penalizes structural negative correlation. The factorial scheme 
 discriminates more strongly the blocks than the centroid one.}

\item{superblock}{Boolean indicating the presence of the superblock. 
Default = TRUE}

\item{type}{A character string indicating the type of multi-block component
method to consider: rgcca, sgcca, pca, pls, cca, ifa, ra, cpca-w, gcca, 
hpca, maxbet-b, maxbet, maxdiff-b, maxdiff, maxvar-a, maxvar-b, maxvar, 
niles, r-maxvar, rcon-pca, ridge-gca, sabscor, ssqcor, ssqcor, ssqcov-1, 
ssqcov-2, ssqcov,  sum-pca, sumcor, sumcov-1, sumcov-2, sumcov, sabscov, 
plspm.}

\item{verbose}{A boolean displaying the warnings}

\item{quiet}{A boolean hidding the warnings}

\item{response}{An integer giving the position of the response block within 
the blocks (activates the supervised mode).}
}
\value{
\item{blocks}{A list of matrix}

\item{scheme}{A character or a function giving the link function for 
covariance maximization among "horst" (the identity function), "factorial"
 (the squared values), "centroid" (the absolute values). Only, the horst 
 scheme penalizes structural negative correlation. The factorial scheme 
 discriminates more strongly the blocks than the centroid one.}

\item{penalty}{A vector of floats (or character for 'optimal' 
setting) giving the shrinkage parameter for covariance maximization}

\item{ncomp}{A vector of 1*J integers giving the number of component 
for each blocks}

\item{connection}{A symmetric matrix (J*J) that describes the 
relationships between blocks}

\item{superblock}{A boolean giving the presence (TRUE) / absence
(FALSE) of a superblock}
}
\description{
Define the parameters associated with each well-known multi-block component
method
}
