% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_connection.R
\name{set_connection}
\alias{set_connection}
\title{Create either a superblock design matrix (if superblock = TRUE), or a
supervised design matrix (if response != NULL) or a fully connected design
matrix (if response == NULL and superblock == FALSE)}
\usage{
set_connection(blocks, superblock = FALSE, response = NULL)
}
\arguments{
\item{blocks}{A list that contains the J blocks of variables
\eqn{\mathbf{X_1}, \mathbf{X_2}, ..., \mathbf{X_J}}{X1, X2, ..., XJ}.
Block \eqn{\mathbf{X}_j}{Xj} is a matrix of dimension
\eqn{n \times p_j}{n x p_j} where n is the number of
observations and \eqn{p_j} the number of variables.}

\item{superblock}{Boolean indicating the presence of the superblock.
Default = TRUE}

\item{response}{Position of the response block}
}
\value{
A binary design matrix encoding the connection between the blocks
}
\description{
Create either a superblock design matrix (if superblock = TRUE), or a
supervised design matrix (if response != NULL) or a fully connected design
matrix (if response == NULL and superblock == FALSE)
}
