% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_connection.R
\name{set_connection}
\alias{set_connection}
\title{Create a matrix corresponding to a connection between the blocks}
\usage{
set_connection(blocks, superblock = FALSE, response = NULL)
}
\arguments{
\item{blocks}{A list of matrices giving the \eqn{J} blocks of variables \eqn{\mathbf{X_1}, \mathbf{X_2}, ..., \mathbf{X_J}}.}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}

\item{response}{if not NULL, an integer giving the position of the response block within the blocks}
}
\value{
A matrix corresponding to the connection between the blocks
}
\description{
Create a matrix corresponding to a connection between the blocks
}
