% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_connection.R
\name{set_connection}
\alias{set_connection}
\title{Create a matrix corresponding to a connection between the blocks}
\usage{
set_connection(blocks, superblock = FALSE)
}
\arguments{
\item{blocks}{A list of matrix}

\item{superblock}{A boolean giving the presence (TRUE) / absence (FALSE) of
a superblock}
}
\value{
A matrix corresponding to the connection between the blocks
}
\description{
Create a matrix corresponding to a connection between the blocks
}
