% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichNAmeth.r
\name{whichNAmethod}
\alias{whichNAmethod}
\title{comparison of two RGCCA results}
\usage{
whichNAmethod(A, listNAdataset = NULL, C = matrix(1, length(A),
  length(A)) - diag(length(A)), tau = rep(1, length(A)), listMethods,
  nDatasets = 20, patternNA = NULL, typeNA = "block",
  ncomp = rep(2, length(A)), sameBlockWeight = TRUE, scale = TRUE,
  tol = 1e-06, verbose = FALSE, scheme = "centroid", seed = NULL)
}
\arguments{
\item{A}{A list of complete blocks}

\item{listNAdataset}{if TRUE, no RGCCA on complete data is run}

\item{C, tau, scale, scheme, sameBlockWeight, ncomp, tol, verbose}{parameters of RGCCA}

\item{listMethods}{vector containing a list of methods ("mean","complete","nipals"...)}

\item{nDatasets}{Number of simulated dataset}

\item{patternNA}{number of missing data required}

\item{typeNA}{structure of missing data required ("ponc" or "block")}

\item{seed}{if filled (by a number), the randomness is reproducible.}
}
\value{
\item{A}{A list of dataset indicators containg a list of rgcca with a  list of criterion. }

\item{crit}{Convergence criterion : abs(1-obj_k/obj_{k-1})}

\item{obj}{Vector containing the mean square error between the predicted values and the original non missing values at each iteration}
}
\description{
Analysis of the comparison of different NA methods on RGCCA
}
\examples{
set.seed(42);X1=matrix(rnorm(350),70,5);X2=matrix(rnorm(280),70,4);
colnames(X1)=paste("A",1:5);colnames(X2)=paste("B",1:4);
rownames(X1)=rownames(X2)=paste("S",1:70)
A=list(X1,X2);
res=whichNAmethod(A,listMethods=c("nipals","mean"),patternNA=rep(0.1,2))
}
